Lifetime movement history of a potamodromous freshwater fish

How the presence and timing of movements are correlated with annual somatic growth rates is explored, a useful proxy for performance because it is easily measured and rapidly reflects environmental changes, and indicates that lifetime movement is linked to faster growth in the early years of a fish's life.

bt.barrow.2021

Format

The data frame 75,343 × 7 contains the following columns:

fish_id_anon	character	fish identifier

distance	numeric	distance from the core to the edge of otolith (um)

oto_sr8786	numeric	otolith 87Sr/86Sr (permil)

age	integer	fish age (years)

date	numeric	date of capture

capture_location	character	location of capture

length	numeric	fish length (mm)

Details

The dataset contains 87Sr/86Sr values in otoliths of 213 individual golden perch Macquaria ambigua from previous studies that sampled MDB rivers in 2014–2017 (Zampatti et al., 2018; Zampatti, Strawbridge, et al., 2018).

Instrument: LA-ICP-MS (lLaser ablation-inductively coupled plasma mass spectrometry)

Source

Barrow, J. S., Yen, J. D., Koehn, J. D., Zampatti, B. P., Thiem, J. D., Tonkin, Z., ... & Morrongiello, J. R. (2021). Lifetime movement history is associated with variable growth of a potamodromous freshwater fish. Journal of Animal Ecology, 90(11), 2560-2572. https://doi.org/10.1111/1365-2656.13561

Zampatti, B. P., Leigh, S. J., Bice, C. M., & Rogers, P. J. (2018). Multiscale movements of golden perch (Percichthyidae: Macquaria ambigua) in the River Murray, Australia. Austral Ecology, 43(7), 763-774. https://doi.org/10.1111/aec.12619

Zampatti, B. P., Strawbridge, A., Thiem, J. D., Tonkin, Z., Mass, R., Woodhead, J., & Fredberg, J. (2018). Golden perch (Macquaria ambigua) and silver perch (Bidyanus bidyanus) age demographics, natal origin and migration history in the River Murray. S. Publication & S.R.R.S. F2018/000116-1, South Australian Research and Development Institute (Aquatic Sciences).

Data availability are available at https://doi.org/10.26188/14122688

Traversing the paper's information via Semantic Scholar ID b6c28496cf300369595c696010ba6b8f1e7245fc using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, vertebra, trace element, stable isotope, Sr8786, Sr/Ca, d34S

Examples

### copy data into 'dat'
dat <- bt.barrow.2021
tibble::tibble(dat)
#> # A tibble: 75,343 × 7
#>    fish_id_anon distance oto_sr8786   age date       capture_location length
#>    <chr>           <int>      <dbl> <int> <chr>      <chr>             <int>
#>  1 fish_32             0      0.707     7 29/04/2016 Lower.Murray        361
#>  2 fish_33             0      0.708     5 29/04/2016 Lower.Murray        253
#>  3 fish_34             0      0.713    19 29/04/2016 Lower.Murray        362
#>  4 fish_35             0      0.707     6 29/04/2016 Lower.Murray        409
#>  5 fish_36             0      0.713    19 29/04/2016 Lower.Murray        404
#>  6 fish_37             0      0.707     5 29/04/2016 Lower.Murray        239
#>  7 fish_38             0      0.708     5 6/04/2016  Lower.Murray        335
#>  8 fish_39             0      0.713    19 7/04/2016  Lower.Murray        373
#>  9 fish_40             0      0.714    19 12/04/2016 Lower.Murray        379
#> 10 fish_41             0      0.709     6 12/04/2016 Lower.Murray        272
#> # ℹ 75,333 more rows

if (FALSE) {
library(dplyr)
library(ggplot2)

### Sr stable isotope
ggplot(data = dat, aes(distance, oto_sr8786)) +
  geom_line(aes(colour = capture_location, group = fish_id_anon),
            show.legend = F, na.rm = T) +
  facet_grid(capture_location ~. ,scales = "free_y")+
  labs(
    x = expression(paste("Distance from the core to the edge of otolith (", mu, "m)",
                         sep = "")),
    y = expression(paste(
      {}^"87",
      "Sr/",
      {}^"86",
      "Sr"
    ))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}