Quantifying estuarine habitat use of Chinook salmon via otolith analyses

The authors use genetic stock identification and otolith analyses to quantify estuarine habitat use by Chinook salmon (Oncorhynchus tshawytscha)-the Pacific salmon species considered most reliant on this habitat-in Canada's most productive salmon river, the Fraser River.

bt.chalifour.2021

Format

The data frame 91 × 8 contains the following columns:

fishid	integer	fish identifier

habitat	character	habitat type

forklength_mm	numeric	fork length (mm)

side	character	otolith side (L or R)

sr_edge	numeric	sr/ca of otolith edge (mmol/mol)

ba_edge	numeric	ba/ca of otolith edge (umol/mol)

sr_core	numeric	sr/ca of otolith core (mmol/mol)

ba_core	numeric	ba/ca of otolith core (umol/mol)

Details

Sr:Ca at the otolith edge increases, and Ba:Ca at the otolith edge decreases with increasing residency in the estuary. This suggests that tidal fluctuations in the estuary and an ontogenetic shift outward towards more saline habitats influence otolith microchemistry over time as the fish reside in this estuarine environment.

Instrument: LA-ICP-MS (laser ablation-inductively coupled plasma mass spectrometry)

Beam diameter: 30um

Scan speed: 5um/s

Reference materials: NIST611, NIST613, NIST615 (National Institutes of Standards and Technology glass standard)

Source

Chalifour, L., Scott, D. C., MacDuffee, M., Stark, S., Dower, J. F., Beacham, T. D., ... & Baum, J. K. (2021). Chinook salmon exhibit long-term rearing and early marine growth in the Fraser River, British Columbia, a large urban estuary. Canadian Journal of Fisheries and Aquatic Sciences, 78(5), 539-550. https://doi.org/10.1139/cjfas-2020-0247

Sturrock, A. M., Wikert, J. D., Heyne, T., Mesick, C., Hubbard, A. E., Hinkelman, T. M., ... & Johnson, R. C. (2015). Reconstructing the migratory behavior and long-term survivorship of juvenile Chinook salmon under contrasting hydrologic regimes. PLoS One, 10(5), e0122380.

Code and data availability are available at https://github.com/baumlab/Chalifour_etal_2021_CJFAS

Traversing the paper's information via Semantic Scholar ID a9b228af81e0e3ac67d0f621adbc10fc55b1fb1a using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, bulk elements, Sr/Ca, Ba/Ca

Examples

### copy data into 'dat'
dat <- bt.chalifour.2021
tibble::tibble(dat)
#> # A tibble: 91 × 8
#>    fishid habitat   forklength_mm side  sr_edge ba_edge sr_core ba_core
#>     <int> <chr>             <int> <chr>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1    762 Marsh                40 L       0.867   1.74    1.30    0.133
#>  2    767 Marsh                44 L       1.55    1.08    1.33    0.677
#>  3    933 Marsh                42 L       1.05    1.47    1.67    0.161
#>  4   4973 Marsh                63 L       0.775   1.71    0.827   0.522
#>  5   4985 Marsh                48 L       1.84    0.294   1.35    0.208
#>  6   4990 Sand flat            71 L       2.27    0.200   1.21    0.521
#>  7   4991 Sand flat            63 L       0.950   0.685   1.21    0.441
#>  8   4992 Sand flat            58 L       1.19    0.735   1.39    0.599
#>  9   4993 Sand flat            67 L       1.14    0.664   1.38    0.526
#> 10   5223 Sand flat            68 R       1.81    0.660   1.40    0.423
#> # … with 81 more rows

if (FALSE) {
### loading packages
library(dplyr)
library(ggplot2)

### otolith core
ggplot(data = dat, aes(sr_core, ba_core)) +
  geom_point(aes(col = habitat), show.legend = F, na.rm = T) +
  facet_grid(. ~ habitat) +
  xlab("Sr/Ca of otolith core (mmol/mol)") +
  ylab("Ba/Ca of otolith core (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### otolith edge
ggplot(data = dat, aes(sr_edge, ba_edge)) +
  geom_point(aes(col = habitat), show.legend = F, na.rm = T) +
  facet_grid(. ~ habitat) +
  xlab("Sr/Ca of otolith edge (mmol/mol)") +
  ylab("Ba/Ca of otolith edge (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}