Studies of spring-run chinook salmon life history diversity

Results from Chinook salmon otolith strontium isotope profiles separated into early, intermediate and late migrants.

bt.cordoleani.2021

Format

The data frame 2,969 × 9 contains the following columns:

sample	character	fish id

year	integer	year of captured

watershed	character	watershed of captured

distance	numeric	distance of from otolith core to edge

otoSr	numeric	otolith strontium isotope data (Sr8786)

SE1	numeric	standard error of otolith strontium isotope data (Sr8786)

SrV	numeric	srv

SrVcol	numeric	srvcol

reartype	character	three life history types of chinook salmon

Details

The dataset contains Chinook salmon otolith strontium isotope profiles. Otolith isotope profiles revealed three distinct juvenile life history types (hereafter referred to as 'early', 'intermediate' and 'late' migrants.

In threatened spring-run Chinook salmon spawning at the southern edge of the species range, this study found that late-migrating juveniles are critical to cohort success in years characterized by droughts and ocean heatwaves. Late migrants rely on cool river temperatures over summer, increasingly rare due to the combined effects of warming and impassable dams.

Source

Cordoleani, F., Phillis, C. C., Sturrock, A. M., FitzGerald, A. M., Malkassian, A., Whitman, G. E., ... & Johnson, R. C. (2021). Threatened salmon rely on a rare life history strategy in a warming landscape. Nature Climate Change, 11(11), 982--988. https://doi.org/10.1038/s41558-021-01186-4

Data and code availability are available at https://github.com/floracordoleani/MillDeerOtolithPaper

Traversing the paper's information via Semantic Scholar ID e71577a8327b918628b43a1964ef212ee802e9a4 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, isotopic ratios, Sr8786

Examples

### copy data into 'dat'
dat <- bt.cordoleani.2021
tibble::tibble(dat)
#> # A tibble: 2,969 × 9
#>    sample   year watershed  distance otoSr     SE1   SrV SrVcol reartype              
#>    <chr>   <int> <chr>         <dbl> <dbl>   <dbl> <dbl>  <dbl> <chr>                 
#>  1 DC07_01  2007 Deer Creek     26.3 0.707 0.00009  1.02  41.3  IntermediateOutmigrant
#>  2 DC07_01  2007 Deer Creek     82.5 0.707 0.00008  0.91  23.8  IntermediateOutmigrant
#>  3 DC07_01  2007 Deer Creek    150.  0.706 0.00012  0.92  25.4  IntermediateOutmigrant
#>  4 DC07_01  2007 Deer Creek    203.  0.705 0.00013  0.76   0    IntermediateOutmigrant
#>  5 DC07_01  2007 Deer Creek    268.  0.705 0.00012  0.93  27.0  IntermediateOutmigrant
#>  6 DC07_01  2007 Deer Creek    293.  0.704 0.00012  1.13  58.7  IntermediateOutmigrant
#>  7 DC07_01  2007 Deer Creek    328.  0.704 0.00012  0.8    6.35 IntermediateOutmigrant
#>  8 DC07_01  2007 Deer Creek    358.  0.704 0.00012  0.87  17.5  IntermediateOutmigrant
#>  9 DC07_01  2007 Deer Creek    387.  0.705 0.00008  0.79   4.76 IntermediateOutmigrant
#> 10 DC07_01  2007 Deer Creek    408.  0.704 0.00014  0.78   3.17 IntermediateOutmigrant
#> # … with 2,959 more rows

if (FALSE) {
### Sr profile figure
library(dplyr)
library(ggplot2)

ggplot(data = dat,aes(distance, otoSr))+
  geom_line(aes(col = reartype),show.legend = F,linewidth = 0.02, na.rm = T)+
  facet_grid(.~ reartype)+
  labs(
    x = (expression(paste("Otolith radius (", mu, "m)", sep = ""))),
    y = expression(paste(
      {}^"87",
      "Sr/",
      {}^"86",
      "Sr"
    ))
  )+
  scale_x_continuous(limits = c(0, 1200), breaks = seq(0, 1200, 200)) +
  scale_y_continuous(limits = c(0.7035, 0.710), breaks = seq(0.704, 0.710, 0.001)) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

}