Reconstructing habitats occupied of Clear Lake Hitch

The goal of this study was to determine if otolith chemistry could be used to reconstruct the habitats occupied by individual Clear Lake Hitch. Study objectives included determining the relative importance of natal habitats and early life migration histories that have contributed to the production and adult recruitment of Clear Lake Hitch.

bt.feyrer.2019

Format

The data frame 2,007 × 11 contains the following columns:

fish_id	character	fish identifier

distance	numeric	distance from the core to the edge of otolith (um)

sr8786	numeric	otolith 87Sr/86Sr (permil)

sr8786_norm	numeric	otolith sr8786 norm

standar_error	numeric	standar error of sr8786

sr_v	numeric	sr v

date	numeric	date of capture

lattitude	numeric	lattitude (NAD83)

longitude	numeric	longitude (NAD83)

length	numeric	fish length (mm)

weight	character	body weight (g)

Details

The dataset includes four seperate files of data for a project that examined the strontium isotopic composition of otoliths of 46 Clear Lake Hitch Lavinia exilicauda chi. Otoliths examined in this study were from randomly-sampled individual Clear Lake Hitch that recruited into the adult population in Clear Lake. The fish were collected in June and July 2017. Strontium isotopic composition (87Sr/86Sr ) in otoliths was determined with a multicollector LA-ICP mass spectrometer (LA-ICMPS; Nu Plasma HR, Nu Instrument, Inc.) interfaced with a Nd : YAG 213-nm laser (New Wave Research UP213) at the University of California, Davis.

Instrument: LA-MC-ICP-MS (laser ablation–multicollector–inductively coupled plasma–mass spectrometry)

Beam diameter: 40um

Reference materials: NIST-987 (National Institutes of Standards and Technology)

Source

Feyrer, F., Whitman, G., Young, M., & Johnson, R. C. (2019). Strontium isotopes reveal ephemeral streams used for spawning and rearing by an imperilled potamodromous cyprinid Clear Lake hitch Lavinia exilicauda chi. Marine and Freshwater Research, 70(12), 1689-1697. https://doi.org/10.1071/MF18264

Data availability are available at https://doi.org/10.5066/P9IX7L5V

Traversing the paper's information via Semantic Scholar ID 19fc02676d9a25c4a9896dc543d17f716ccb8c3c using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, isotopic ratios, Sr8786

Examples

### copy data into 'dat'
dat <- bt.feyrer.2019
tibble::tibble(dat)
#> # A tibble: 2,007 × 11
#>    fish_id      distance sr8786 sr8786_norm standar_error  sr_v date      lattitude longitude length weight
#>    <chr>           <dbl>  <dbl>       <dbl>         <dbl> <dbl> <chr>         <dbl>     <dbl>  <int>  <int>
#>  1 CLH_2017_001     9.97  0.707       0.707     0.000117   3.03 7/20/2017      39.0     -123.    280    425
#>  2 CLH_2017_001    30.4   0.706       0.707     0.0000649  3.52 7/20/2017      39.0     -123.    280    425
#>  3 CLH_2017_001    42.1   0.707       0.707     0.0000982  3.10 7/20/2017      39.0     -123.    280    425
#>  4 CLH_2017_001    65.7   0.706       0.706     0.0000614  4.52 7/20/2017      39.0     -123.    280    425
#>  5 CLH_2017_001    67.8   0.706       0.706     0.0000598  4.30 7/20/2017      39.0     -123.    280    425
#>  6 CLH_2017_001   113.    0.706       0.706     0.000104   4.47 7/20/2017      39.0     -123.    280    425
#>  7 CLH_2017_001   156.    0.706       0.706     0.000077   4.23 7/20/2017      39.0     -123.    280    425
#>  8 CLH_2017_001   191.    0.706       0.706     0.0000489  4.26 7/20/2017      39.0     -123.    280    425
#>  9 CLH_2017_001   234.    0.706       0.706     0.0000688  4.11 7/20/2017      39.0     -123.    280    425
#> 10 CLH_2017_001   273.    0.706       0.706     0.0000814  4.03 7/20/2017      39.0     -123.    280    425
#> # … with 1,997 more rows

if (FALSE) {
library(dplyr)
library(ggplot2)

### Sr stable isotope
ggplot(data = dat, aes(distance,sr8786)) +
  geom_line(aes(group = fish_id),show.legend = F, na.rm = T, colour = "blue") +
  labs(
    x = expression(paste("Distance from the core to the edge of otolith (", mu, "m)",
                         sep = "")),
    y = expression(paste(
      {}^"87",
      "Sr/",
      {}^"86",
      "Sr"
    ))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}