Otolith chemistry chronicles life histories of coastal lagoon fishes

This study examined the habitat ecology of two lagoon species, the endangered Tidewater Goby (Eucyclogobius newberryi) and the Prickly Sculpin (Cottus asper) by reconstructing individual life histories from patterns in the concentration of the element Sr (as ratioed to Ca; Sr:Ca) in otoliths.Specific objectives were to (1) elucidate any movements of individual fishes among three primary habitat components of typical bar-built lagoon systems: coastal ocean, brackish lagoon, and freshwater watershed streams, and (2) determine if either species exhibited a standard life history as defined by a prototypical otolith Sr:Ca chronology, which could be indicative of a standard range of salinity or temperature occupied through ontogeny.

bt.feyrer.2020

Format

The data frame 5,734 × 11 contains the following columns:

fish_id	integer	fish identifier

distance	numeric	distance from the core to the edge of otolith (microns)

value	numeric	Sr:Ca or Mn:Ca

element	character	trace element

common_name	character	species common name

scientific_name	character	species scientific name

location	character	location of capture

date	character	date of capture

standard_length	numeric	standard length (mm)

weight	numeric	fish weight (g)

otolith_radius	numeric	otolith_radius (microns)

Details

This dataset contains otolith elemental raw data of two coastal lagoon fishes from around the perimeter of Rodeo Lagoon following a hypoxia-induced fish kill that occurred on 08 August 2016. Additional Prickly Sculpins were collected on 12 April 2016 from Rodeo Lagoon using a beach seine and from Rodeo Creek using a minnow trap. In total, 14 Tidewater Gobies (mean standard length = 40 mm, standard deviation = 2) and 10 Prickly Sculpins (mean standard length = 62 mm, standard deviation = 12) were examined.

Instrument: SXFM (scanning x-ray fluorescence imaging)

Source

Feyrer, F., Young, M., Fong, D., Limburg, K., & Johnson, R. (2020). Cryptic Lives of Conspicuous Animals: Otolith Chemistry Chronicles Life Histories of Coastal Lagoon Fishes. Frontiers in Marine Science, 7, 417. https://doi.org/10.3389/fmars.2020.00417

Data availability are available at https://doi.org/10.5066/P9PZMELL

Traversing the paper's information via Semantic Scholar ID 57c668a6b12ca55f2534f448db15ddbe8212c2f9 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, trace element, Sr/Ca, Mn/Ca

Examples

### copy data into 'dat'
dat <- bt.feyrer.2020
tibble::tibble(dat)
#> # A tibble: 5,734 × 11
#>    fish_id distance   value element common_name  scientific_name location date  standard_length weight
#>      <int>    <int>   <dbl> <chr>   <chr>        <chr>           <chr>    <chr>           <int>  <dbl>
#>  1       6        0 0.00389 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#>  2       6        1 0.00395 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#>  3       6        2 0.00393 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#>  4       6        3 0.00393 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#>  5       6        4 0.00395 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#>  6       6        5 0.00405 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#>  7       6        6 0.00402 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#>  8       6        7 0.00398 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#>  9       6        8 0.00394 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#> 10       6        9 0.00411 Sr_Ca   Tidewater G… " Eucyclogobiu… Rodeo L… 8/16…              38   0.64
#> # ℹ 5,724 more rows
#> # ℹ 1 more variable: otolith_radius <dbl>

if (FALSE) {
### load package
library(dplyr)
library(ggplot2)

### otolith sr/ca
ggplot(data = dat[which(dat$element == "Sr_Ca" & dat$value < 0.02), ],
       aes(distance, value)) +
  geom_line(aes(colour = common_name, group = fish_id), show.legend = F, na.rm = T) +
  facet_grid(common_name ~ ., scales = "free_y") +
  xlab("Distance from the core to the edge of otolith (mm)") +
  ylab("Otolith Sr:Ca") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
### otolith mn/ca
ggplot(data = dat[which(dat$element == "Mn_Ca" & dat$value < 0.001),],
       aes(distance, value)) +
  geom_line(aes(colour = common_name, group = fish_id), show.legend = F, na.rm = T) +
  facet_grid(common_name ~ ., scales = "free_y") +
  xlab("Distance from the core to the edge of otolith (mm)") +
  ylab("Otolith Mn:Ca") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}