Coupling biogeochemical tracers with fish growth

The lack of among-individual correlations of otolith traits in properly formulated models and the biases that can be introduced by not including appropriate intrinsic covariates suggest that caution is needed when assuming multi-elemental signatures are reflective solely of shared environments.

bt.grammer.2017

Format

The data frame 2,688 × 14 contains the following columns:

id_code	character	unique identifier for each fish

age_years_cap	integer	age of the fish in years at date of capture

age_month_cap	integer	age of the fish in months at date of capture

month_code_cap	integer	month code for the date of capture, e.g. 75 = Dec 1996

fish_id	character	= id_code

month	integer	a code designating the specific month and year of each monthly resolved increment (i.e. 75 = Dec 1996, 76 = Jan 1997, etc.)

age	integer	age when each monthly increment was formed (total months)

f_month	integer	= month (as factor)

growth	numeric	measured value of the growth increment (μm) for that discrete monthly time period

sr_ca	numeric	value of the Sr:Ca ratio for that discrete monthly time period (μmol/mol)

na_ca	numeric	value of the Na:Ca ratio for that discrete monthly time period (μmol/mol)

ba_ca	numeric	value of the Ba:Ca ratio for that discrete monthly time period (μmol/mol)

mg_ca	numeric	value of the Mg:Ca ratio for that discrete monthly time period (μmol/mol)

li_ca	numeric	value of the Li:Ca ratio for that discrete monthly time period (μmol/mol)

Details

Value of the Element:Ca ratio for that discrete monthly time period (μmol/mol) extracted from the laser ablation ICP-MS data.

Instrument: Q-ICP-MS (quadrupole inductively coupled plasma mass spectrometry)

Beam diameter: 33um

Scan speed: 4um/s

Reference materials: NIST-612 (National Institutes of Standards and Technology glass standard) and MACS-3 (United States Geological Survey microanalytical carbonate standard pressed pellet)

Source

Grammer, G. L., Morrongiello, J. R., Izzo, C., Hawthorne, P. J., Middleton, J. F., & Gillanders, B. M. (2017). Coupling biogeochemical tracers with fish growth reveals physiological and environmental controls on otolith chemistry. Ecological Monographs, 87(3), 487-507. https://doi.org/10.1002/ecm.1264

Code and data availability are available at https://doi.org/10.5061/dryad.cn55b

Traversing the paper's information via Semantic Scholar ID 8255e6b7bc33ad3ba95ccc7313846c631539c773 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, bulk elements, Sr/Ca, Na/Ca, Ba/Ca, Mg/Ca, Li/Ca

Examples

### copy data into 'dat'
dat <- bt.grammer.2017
tibble::tibble(dat)
#> # A tibble: 2,688 × 14
#>    id_code age_years_cap age_month…¹ month…² fish_id month   age f_month growth sr_ca  na_ca ba_ca mg_ca li_ca
#>    <chr>           <int>       <int>   <int> <chr>   <int> <int>   <int>  <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 HPS.004            15         181     253 HPS.004    86    14      86   25.7 2968. 13073.  2.64  63.9 11.5 
#>  2 HPS.004            15         181     253 HPS.004    95    23      95   24.1 2933. 13944.  1.53  86.4  7.72
#>  3 HPS.004            15         181     253 HPS.004    96    24      96   22.2 3175. 12527.  1.66  65.3  6.86
#>  4 HPS.004            15         181     253 HPS.004    97    25      97   21.4 3268. 11409.  1.52  49.7  6.52
#>  5 HPS.004            15         181     253 HPS.004    98    26      98   17.6 3338. 11384.  1.53  44.9  6.13
#>  6 HPS.004            15         181     253 HPS.004    99    27      99   17.5 3299. 11993.  1.34  43.4  7.93
#>  7 HPS.004            15         181     253 HPS.004   100    28     100   15.6 2888. 11756.  1.16  43.1  7.90
#>  8 HPS.004            15         181     253 HPS.004   101    29     101   14.8 2892. 12946.  1.14  53.9  6.87
#>  9 HPS.004            15         181     253 HPS.004   102    30     102   14.8 2852. 13550.  1.14  52.7  6.82
#> 10 HPS.004            15         181     253 HPS.004   103    31     103   14.0 2829. 13177.  1.13  49.8  7.11
#> # … with 2,678 more rows, and abbreviated variable names ¹​age_month_cap, ²​month_code_cap

if (FALSE) {

library(dplyr)
library(ggplot2)

### Sr/Ca
ggplot(data = dat, aes(age, sr_ca)) +
  geom_line(aes(group = fish_id), show.legend = F, na.rm = T) +
  labs(
    x = expression("Age (months)"),
    y = expression(paste("Sr/Ca (", μmol / mol, ")", sep = ""))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Na/Ca
ggplot(data = dat, aes(age, na_ca)) +
  geom_line(aes(group = fish_id), show.legend = F, na.rm = T) +
  labs(
    x = expression("Age (months)"),
    y = expression(paste("Na/Ca (", μmol / mol, ")", sep = ""))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Ba/Ca
ggplot(data = dat, aes(age, ba_ca)) +
  geom_line(aes(group = fish_id), show.legend = F, na.rm = T) +
  labs(
    x = expression("Age (months)"),
    y = expression(paste("Ba/Ca (", μmol / mol, ")", sep = ""))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Mg/Ca
ggplot(data = dat, aes(age, mg_ca)) +
  geom_line(aes(group = fish_id), show.legend = F, na.rm = T) +
  labs(
    x = expression("Age (months)"),
    y = expression(paste("Mg/Ca (", μmol / mol, ")", sep = ""))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Li/Ca
ggplot(data = dat, aes(age, li_ca)) +
  geom_line(aes(group = fish_id), show.legend = F, na.rm = T) +
  labs(
    x = expression("Age (months)"),
    y = expression(paste("Li/Ca (", μmol / mol, ")", sep = ""))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Growth

ggplot(data = dat, aes(age, growth)) +
  geom_line(aes(group = fish_id), show.legend = F, na.rm = T) +
  labs(
    x = expression("Age (months)"),
    y = expression(paste("Growth (", mu, "m)", sep = ""))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}