Unravelling the life history of Amazonian fishes

The results of this study suggested that Brachyplatystoma filamentosum did not enter the estuary during its life history, and introduced what should be a productive line of research desperately needed to better understand the migrations of these unique and imperilled fishes.

bt.hermann.2016

Format

The data frame 7,035 × 6 contains the following columns:

sampleid	character	fish identifier

species	character	species studied

distance	numeric	distance from fish otolith cores-edge transects (unit: micron)

value	character	Laser ablation–inductively coupled plasma mass spectrometry raw value

oto_element	character	bulk elements: Ba, Sr, Ca, Mn

locality	character	locality of captured

Details

Laser ablation–inductively coupled plasma mass spectrometry raw data (background- and drift-corrected) for Brachyplatystoma rousseauxii (Ecuador and Colombia), Brachyplatystoma filamentosum (Colombia), and six Prochilodus nigricans. Data include Ca, Sr, and Ba for both Brachyplatystoma species and localities, and Ca, Ba, and Mn for all six P. nigricans.

Source

Hermann, T. W., Stewart, D. J., Limburg, K. E., & Castello, L. (2016). Unravelling the life history of Amazonian fishes through otolith microchemistry. Royal Society open science, 3(6), 160206. http://dx.doi.org/10.1098/rsos.160206

Data availability are available at https://doi.org/10.5061/dryad.8kd35

Traversing the paper's information via Semantic Scholar ID 73c934cff5baaa29235d0388b4f14133c95415f7 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, bulk elements, Sr/Ca, Mn/Ca, Ba/Ca

Examples

### copy data into 'dat'
dat <- bt.hermann.2016
tibble::tibble(dat)
#> # A tibble: 7,035 × 6
#>    sampleid species                      distance    value oto_element locality
#>    <chr>    <chr>                           <dbl>    <dbl> <chr>       <chr>   
#>  1 BR1      Brachyplatystoma rousseauxii     0     487371. Ca          Ecuador 
#>  2 BR1      Brachyplatystoma rousseauxii     7.59  827141. Ca          Ecuador 
#>  3 BR1      Brachyplatystoma rousseauxii    15.2  1025308  Ca          Ecuador 
#>  4 BR1      Brachyplatystoma rousseauxii    22.8   476471. Ca          Ecuador 
#>  5 BR1      Brachyplatystoma rousseauxii    30.4   860977. Ca          Ecuador 
#>  6 BR1      Brachyplatystoma rousseauxii    37.9   638388  Ca          Ecuador 
#>  7 BR1      Brachyplatystoma rousseauxii    45.5   665656. Ca          Ecuador 
#>  8 BR1      Brachyplatystoma rousseauxii    53.1   609487. Ca          Ecuador 
#>  9 BR1      Brachyplatystoma rousseauxii    60.7   759480. Ca          Ecuador 
#> 10 BR1      Brachyplatystoma rousseauxii    68.3   509173. Ca          Ecuador 
#> # … with 7,025 more rows
table(dat$species)
#> 
#> Brachyplatystoma filamentosum  Brachyplatystoma rousseauxii         Prochilodus nigricans 
#>                          1125                          3612                          2298 

if (FALSE) {
### bulk elements profile figure
library(dplyr)
library(tidyr)
library(ggplot2)

### Brachyplatystoma rousseauxii

dat[which(dat$species == "Brachyplatystoma rousseauxii" &
            dat$locality == "Ecuador"), ]
B_rousseauxii |>
  pivot_wider(
    names_from = oto_element,
    values_from = value
  ) |>
  ggplot(aes(distance, (Ba / Ca) * 1000)) +
  geom_line(aes(col = locality), show.legend = F, linewidth = 0.02, na.rm = T) +
  labs(
    x = (expression(
      paste("Distance from fish otolith cores – edge transects (", mu, "m)",
            sep = ""))),
    y = ("(Ba/Ca)x1000")
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

B_rousseauxii |>
  pivot_wider(
    names_from = oto_element,
    values_from = value
  ) |>
  ggplot(aes(distance, (Sr / Ca) * 1000)) +
  geom_line(aes(col = locality), show.legend = F, linewidth = 0.02, na.rm = T) +
  labs(
    x = (expression(
      paste("distance from fish otolith cores – edge transects (", mu, "m)",
            sep = ""))),
    y = ("Sr/Ca")
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}