Investigating natal philopatry in Hickory Shad using otolith chemistry

LA-ICP-MS was used to quantify seven elements (Mg, Mn, Cu, Zn, Sr, Ba, and Pb), along a continuous transect that ran from the ventral edge of the otolith through the core to the dorsal edge of each otolith, resulting in a time resolved model of the environmental exposure history of each fish.

bt.hill.2020

Format

The data frame 266,731 × 16 contains the following columns:

river	character	river of capture

parent_river	character	parent river of capture

estuary	character	estuary of capture

state	character	state of capture

capt_yr	integer	year of capture

fishno	integer	fish id

sex	character	fish sex

elapsed_time	numeric	elapsed time of LA-ICP-MS

distance	numeric	distance from the ventral edge of the otolith through the core to the dorsal edge

mg_ca	numeric	Mg/Ca

mn_ca	numeric	Mn/Ca

cu_ca	numeric	Cu/Ca

zn_ca	numeric	Zn/Ca

sr_ca	numeric	Sr/Ca

ba_ca	numeric	Ba/Ca

pb_ca	numeric	Pb/Ca

Details

The dataset contains contains the raw chemistry data (results of LA-ICP-MS) for all 289 Hickory Shad (Alosa mediocris).

Instrument: LA-ICP-MS (laser ablation-inductively coupled plasma mass spectrometry)

Beam diameter: 30um

Scan speed: 2-5um/s

Reference materials: NIST-610 (National Institutes of Standards and Technology glass standard)

Source

Hill, C. R., Rulifson, R. A., & Halden, N. M. (2022). A quantitative approach to investigate natal philopatry in Hickory Shad (Alosa mediocris) using otolith chemistry. Fisheries Research, 249, 106232. https://doi.org/10.1016/j.fishres.2022.106232

Code and data availability are available at https://doi.org/10.5281/zenodo.4536534

Traversing the paper's information via Semantic Scholar ID 7690935e5be9a70157633875a316d7ba3d083205 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, trace element, Sr/Ca, Mn/Ca, Mg/Ca, Ba/Ca, Pb/Ca, Cu/Ca

Examples

### copy data into 'dat'
dat <- bt.hill.2020
tibble::tibble(dat)
#> # A tibble: 266,731 × 16
#>    river parent_river estuary state capt_yr fishno sex   elapsed_time distance mg_ca mn_ca cu_ca zn_ca
#>    <chr> <chr>        <chr>   <chr>   <int>  <int> <chr>        <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Alt   Alt          AltSd   GA       2017    786 F              0        0    0.05  0.43  0.63  0.04
#>  2 Alt   Alt          AltSd   GA       2017    786 F              0.8      2.4  0.04  0.91  0.8   0.38
#>  3 Alt   Alt          AltSd   GA       2017    786 F              1.6      4.8  0.05  0.5   0.4   1   
#>  4 Alt   Alt          AltSd   GA       2017    786 F              2.4      7.2  0.03  0.27  0.5  -0.03
#>  5 Alt   Alt          AltSd   GA       2017    786 F              3.2      9.6  0.04  0.39  0.39  0.44
#>  6 Alt   Alt          AltSd   GA       2017    786 F              4       12    0.02  0.57  0.18 -0.16
#>  7 Alt   Alt          AltSd   GA       2017    786 F              4.8     14.4  0.03  0.66  0.54  0.38
#>  8 Alt   Alt          AltSd   GA       2017    786 F              5.6     16.8  0.02  0.64  0.14  0.75
#>  9 Alt   Alt          AltSd   GA       2017    786 F              6.4     19.2  0.03  0.8   0.1  -0.13
#> 10 Alt   Alt          AltSd   GA       2017    786 F              7.2     21.6  0.03  0.37  0.43  0.16
#> # ℹ 266,721 more rows
#> # ℹ 3 more variables: sr_ca <dbl>, ba_ca <dbl>, pb_ca <dbl>

if (FALSE) {
### loading packages
library(dplyr)
library(ggplot2)

### Sr/Ca profile figure
ggplot(data = dat, aes(distance, sr_ca)) +
  geom_line(aes(col = sex,group = fishno), show.legend = F) +
  facet_grid(river ~ sex, scales = "free_y") +
  ylab("Sr/Ca")+
  xlab("Distance from the ventral edge of the otolith through the core to the dorsal edge") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Mg/Ca profile figure
ggplot(data = dat, aes(distance, mg_ca)) +
  geom_line(aes(col = sex,group = fishno), show.legend = F) +
  facet_grid(river ~ sex, scales = "free_y") +
  ylab("Mg/Ca")+
  xlab("Distance from the ventral edge of the otolith through the core to the dorsal edge") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Cu/Ca profile figure
ggplot(data = dat, aes(distance, cu_ca)) +
  geom_line(aes(col = sex,group = fishno), show.legend = F) +
  facet_grid(river ~ sex, scales = "free_y") +
  ylab("Cu/Ca")+
  xlab("Distance from the ventral edge of the otolith through the core to the dorsal edge") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Zn/Ca profile figure
ggplot(data = dat, aes(distance, zn_ca)) +
  geom_line(aes(col = sex,group = fishno), show.legend = F) +
  facet_grid(river ~ sex, scales = "free_y") +
  ylab("Zn/Ca")+
  xlab("Distance from the ventral edge of the otolith through the core to the dorsal edge") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Ba/Ca profile figure
ggplot(data = dat, aes(distance, ba_ca)) +
  geom_line(aes(col = sex,group = fishno), show.legend = F) +
  facet_grid(river ~ sex, scales = "free_y") +
  ylab("Ba/Ca")+
  xlab("Distance from the ventral edge of the otolith through the core to the dorsal edge") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Pb/Ca profile figure
ggplot(data = dat, aes(distance, pb_ca)) +
  geom_line(aes(col = sex,group = fishno), show.legend = F) +
  facet_grid(river ~ sex, scales = "free_y") +
  ylab("Pb/Ca")+
  xlab("Distance from the ventral edge of the otolith through the core to the dorsal edge") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}