Low interbasin connectivity in a facultatively diadromous fish

Analysis of multielemental signatures in the otolith cores of diadromous fish revealed strong discrimination between river basins, suggesting that diadromeous fish spend their early lives within chemically distinct estuaries rather than the more homogenous marine environment, thus avoiding dispersal and maintaining genetic structure.

bt.hughes.2014

Format

The data frame 8,834 × 4 contains the following columns:

fish_code	integer	fish identifier

distance	numeric	distance from otolith core (um)

oto_sr8786	numeric	otolith strontium isotope (Sr87/Sr86) values

river	character	river of captured

Details

The datasets contains Strontium isotope ratio data for each individual fish. The authors use otolith chemistry to distinguish between diadromous and nondiadromous life histories and population genetics to examine interbasin connectivity resulting from diadromy. Otolith strontium isotope (87Sr/86Sr) transects identified three main life history patterns: amphidromy, freshwater residency and estuarine/marine residency.

Source

Hughes, J. M., Schmidt, D. J., Macdonald, J. I., Huey, J. A., & Crook, D. A. (2014). Low interbasin connectivity in a facultatively diadromous fish: evidence from genetics and otolith chemistry. Molecular Ecology, 23(5), 1000--1013. http://dx.doi.org/10.1111/mec.12661

Data availability are available at https://doi.org/10.5061/dryad.3dh2c

Traversing the paper's information via Semantic Scholar ID d1015ceaa5cf80e981733e01ca8beb0eb56616b0 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, isotopic ratios, Sr8786

Examples

### copy data into 'dat'
dat <- bt.hughes.2014
tibble::tibble(dat)
#> # A tibble: 8,834 × 4
#>    fish_code distance oto_sr8786 river  
#>        <int>    <dbl>      <dbl> <chr>  
#>  1      3065      0        0.709 Glenelg
#>  2      3065      6.6      0.710 Glenelg
#>  3      3065     13.2      0.710 Glenelg
#>  4      3065     19.8      0.710 Glenelg
#>  5      3065     26.4      0.710 Glenelg
#>  6      3065     33        0.710 Glenelg
#>  7      3065     39.6      0.710 Glenelg
#>  8      3065     46.2      0.710 Glenelg
#>  9      3065     52.8      0.710 Glenelg
#> 10      3065     59.4      0.710 Glenelg
#> # … with 8,824 more rows

if (FALSE) {
### loading packages
library(dplyr)
library(ggplot2)

### Sr8786 profile figure
ggplot(data = dat,aes(distance, oto_sr8786))+
  geom_line(aes(col = river),show.legend = F,linewidth = 0.02, na.rm = T)+
  facet_wrap(.~ river,ncol = 3)+
  labs(
    x = (expression(paste("Distance from otolith core (", mu, "m)", sep = ""))),
    y = (expression(paste(
      {}^"87",
      "Sr/",
      {}^"86",
      "Sr"
    )))
  )+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}