Reconstructing the stable isotope chronology of teleost fishes

It is suggested that the method presented can provide reproducible isotopic chronology, even in teleost fishes smaller than 50 cm, and can be used in isoscape studies and in studies of the ecology of marineteleost fishes.

bt.matsubayashi.2017

Format

The data frame 86 × 8 contains the following columns:

fish_samples	character	fish sample identifier

biomineral_archives	character	biomineral archives (vertebra and otolith)

biogeochemical_tags	character	biogeochemical tags

element_ratio	character	element ratio

section	character	biomineral archives of fish equally into sections

distance	numeric	distance from biomineral archives core to edge

mean_values	numeric	mean values (ug/mg in bulk elements; permil in isotopic ratios)

sd	numeric	standard deviation (ug/mg in bulk elements; permil in isotopic ratios)

Details

Sulfur stable isotope ratios of vertebral sections of masu salmon. Each vertebral centrum was subdivided equally into 10 sections numbered from the center (01) to the margin (10). Fish samples were collected in the Churui River, Hokkaido Islands, japan. All fish samples analyzed consisted of bone collagen.

Elemental and 87Sr/86Sr analyses of fish otolith sections from masu salmon (identified as OM-01, OM-02 and OM-03) collected from the Churui River.

Instrument: TIMS (thermal ionization mass spectrometry);Q-ICP-MS (quadrupole inductively coupled plasma mass spectrometry); IRMS (isotope ratio mass spectrometer)

Reference materials: NIST987 (National Institutes of Standards and Technology), XSTC-622 (SPEX SertiPrep, Metuchen, New Jersey, USA) and IAEA-S-1, IAEA-S-2 (International Atomic Energy Agency standards)

Source

Matsubayashi, J., Saitoh, Y., Osada, Y., Uehara, Y., Habu, J., Sasaki, T., & Tayasu, I. (2017). Incremental analysis of vertebral centra can reconstruct the stable isotope chronology of teleost fishes. Methods in Ecology and Evolution, 8(12), 1755-1763. https://doi.org/10.1111/2041-210X.12834

Data availability are available at https://doi.org/10.5061/dryad.r6p27

Traversing the paper's information via Semantic Scholar ID f0f6b76481b0b6b97d802e49e610b290c30c6e76 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, vertebra, bulk elements, isotopic ratios, Sr8786, Sr/Ca, d34S

Examples

### copy data into 'dat'
dat <- bt.matsubayashi.2017
tibble::tibble(dat)
#> # A tibble: 86 × 8
#>    fish_samples biomineral_archives biogeochemical_tags element_ratio section    distance mean_values    sd
#>    <chr>        <chr>               <chr>               <chr>         <chr>         <int>       <dbl> <dbl>
#>  1 OM-01        vertebra            isotopic ratios     d34S          Section-01        1        12.7  0.41
#>  2 OM-01        vertebra            isotopic ratios     d34S          Section-02        2        13.9  0.41
#>  3 OM-01        vertebra            isotopic ratios     d34S          Section-03        3        17.2  0.41
#>  4 OM-01        vertebra            isotopic ratios     d34S          Section-04        4        15.5  0.41
#>  5 OM-01        vertebra            isotopic ratios     d34S          Section-05        5        17.5  0.41
#>  6 OM-01        vertebra            isotopic ratios     d34S          Section-06        6        18.1  0.41
#>  7 OM-01        vertebra            isotopic ratios     d34S          Section-07        7        18.8  0.41
#>  8 OM-01        vertebra            isotopic ratios     d34S          Section-08        8        18.9  0.41
#>  9 OM-01        vertebra            isotopic ratios     d34S          Section-09        9        19    0.41
#> 10 OM-01        vertebra            isotopic ratios     d34S          Section-10       10        18.9  0.41
#> # … with 76 more rows

if (FALSE) {

library(dplyr)
library(ggplot2)

### Sulfur stable isotope
dat[which(dat$element_ratio == "d34S"), ] |>
  ggplot(aes(distance, mean_values)) +
  geom_point(aes(colour = fish_samples), show.legend = F) +
  geom_line(aes(colour = fish_samples), show.legend = F) +
  labs(
    x = "Vertebral section number",
    y = expression(delta * ""^34 * "S" * " (‰)")
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  ) +
  scale_x_continuous(breaks = 1:10)

### Sr stable isotope
dat[which(dat$element_ratio == "Sr8786"), ] |>
  ggplot(aes(distance, mean_values)) +
  geom_point(aes(colour = fish_samples), show.legend = F) +
  geom_line(aes(colour = fish_samples), show.legend = F) +
  labs(
    x = expression(paste("Distance from core (", mu, "m)", sep = "")),
    y = expression(paste(
      {}^"87",
      "Sr/",
      {}^"86",
      "Sr"
    ))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  ) +
  scale_x_continuous(breaks = c(0, 200, 400, 600, 800, 1000, 1200))

### Sr/Ca
dat[which(dat$element_ratio == "Sr/Ca"), ] |>
  ggplot(aes(distance, mean_values)) +
  geom_point(aes(colour = fish_samples), show.legend = F) +
  geom_line(aes(colour = fish_samples), show.legend = F) +
  labs(
    x = expression(paste("Distance from core (", mu, "m)", sep = "")),
    y = expression(paste("Sr/Ca (", x1000, ")", sep = ""))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  ) +
  scale_x_continuous(breaks = c(0, 200, 400, 600, 800, 1000, 1200))
}