Tracking long-distance migration of marine fishes using stable isotope

The isotope tracking method successfully reproduced a known chum salmon migration route between the Okhotsk and Bering seas, and the findings suggest the presence of a new migration route to the Bering Sea Shelf during a later growth stage.

bt.matsubayashi.2020

Format

The data frame 80 × 11 contains the following columns:

id	character	fish identifier

river	character	river of capture

year	integer	year of capture

month	integer	month of capture

day	integer	day of capture

age	integer	age of capture (years)

fork_length	integer	fork length of capture (mm)

section_no	integer	vertebral bone sections are numbered from the centre of the vertebral centrum and increase toward the margin

d13c	numeric	d13C bulk (permil)

d15n	numeric	d15N bulk (permil)

Details

Specimen IDs and a summary of stable carbon and nitrogen isotope results from the vertebral sections of 8 chum salmon.

Source

Matsubayashi, J., Osada, Y., Tadokoro, K., Abe, Y., Yamaguchi, A., Shirai, K., ... & Tayasu, I. (2020). Tracking long-distance migration of marine fishes using compound-specific stable isotope analysis of amino acids. Ecology Letters, 23(5), 881--890. https://doi.org/10.1111/ele.13496

Data availability are available at https://doi.org/10.5061/dryad.tx95x69tc

Traversing the paper's information via Semantic Scholar ID 32bf4cf7237572a3559a2d0e32912f93c91774a5 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

vertebra, isotopic ratios, d13C, d15N

Examples

### copy data into 'dat'
dat <- bt.matsubayashi.2020
tibble::tibble(dat)
#> # A tibble: 80 × 11
#>    id    river             year month   day sex     age fork_length section_no  d13c  d15n
#>    <chr> <chr>            <int> <int> <int> <chr> <int>       <int>      <int> <dbl> <dbl>
#>  1 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          1 -16.6 10.3 
#>  2 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          2 -16.7 10.4 
#>  3 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          3 -17.1 10.5 
#>  4 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          4 -17.0 10.5 
#>  5 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          5 -17.7  9.38
#>  6 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          6 -18.2  8.85
#>  7 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          7 -17.1  9.77
#>  8 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          8 -18.4  8.65
#>  9 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA          9 -18.5  8.85
#> 10 OK1   Nukkibetsu River  2016    12    29 NA       NA          NA         10 -15.5 12.7 
#> # … with 70 more rows

if (FALSE) {

library(dplyr)
library(ggplot2)

### Carbon stable isotope
ggplot(data = dat, aes(section_no, d13c)) +
  geom_point(aes(colour = id), show.legend = F) +
  geom_line(aes(colour = id), show.legend = F) +
  labs(
    x = "Vertebral section number",
    y = bquote(delta^13 ~C[Bulk]("‰"))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  ) +
  scale_x_continuous(breaks = 1:10)


### Nitrogen stable isotope
ggplot(data = dat, aes(section_no, d15n)) +
  geom_point(aes(colour = id), show.legend = F) +
  geom_line(aes(colour = id), show.legend = F) +
  labs(
    x = "Vertebral section number",
    y = bquote(delta^15 ~N[Bulk]("‰"))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  ) +
  scale_x_continuous(breaks = 1:10)
}