Vertebral sclerochronology in shark*

Vertebral sclerochronology has the potential to advance knowledge of elasmobranch life history including age and growth estimation and environmental reconstruction, and regular oscillations in vertebrae manganese (Mn) content corresponded well with the number of validated band pairs, suggesting that Mn variation could be used to age sharks.

bt.mohan.2018

Format

The data frame 11,118 × 12 contains the following columns:

fish_id	character	shark identifier

vert_sample_id	character	shark vertebral sample identifier

distance	numeric	distance from core to edge of shark vertebrae (um)

value	numeric	vertebral element(E):Ca (Mn/Ca in umol/mol; Ba/Ca in umol/mol; Sr/Ca in mmol/mol)

common_name	character	common name

scientific_name	character	scientific name

recapture_lat	numeric	latitude of recapture

recapture_long	numeric	longitude of recapture

recapture_date	character	date of recapture

recapture_fl	numeric	fork length of recapture (cm)

sex	character	shark sex

Details

The dataset contains element(E):Ca from vertebral core–edge transects of three shark. Elemental profiles were characterized in vertebrae encompassing complete life histories (birth-to-death) of shortfin mako (Isurus oxyrinchus), common thresher (Alopias vulpinus) and blue shark (Prionace glauca) of known tag and recapture locations in the eastern North Pacific Ocean.

Instrument: LA-ICP-MS (laser ablation-inductively coupled plasma mass spectrometry)

Beam diameter: 100-150um

Scan speed: 150um/s

Reference materials: NIST-612 (National Institutes of Standards and Technology glass standard), MACS-3 (United States Geological Survey microanalytical carbonate standard pressed pellet) and MAPS-4

Note: data extraction via WebPlotDigitizer tool (Title with * label).

Source

Mohan, J. A., Miller, N. R., Herzka, S. Z., Sosa-Nishizaki, O., Kohin, S., Dewar, H., ... & Wells, R. D. (2018). Elements of time and place: manganese and barium in shark vertebrae reflect age and upwelling histories. Proceedings of the Royal Society B, 285(1890), 20181760. https://doi.org/10.1098/rspb.2018.1760

Traversing the paper's information via Semantic Scholar ID e65acf2cafe143c21578de2aa2149b3d17ad2cbc using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

vertebrae, bulk elements, Ba/Ca, Mn/Ca, Sr/Ca

Examples

### copy data into 'dat'
dat <- bt.mohan.2018
tibble::tibble(dat)
#> # A tibble: 11,118 × 12
#>    fish_id vert_sample_id distance value element common_name scientifi…¹ recap…² recap…³ recap…⁴ recap…⁵ sex  
#>    <chr>   <chr>             <dbl> <dbl> <chr>   <chr>       <chr>         <dbl>   <dbl> <chr>     <int> <chr>
#>  1 A038423 D1                 34.7  16.3 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#>  2 A038423 D1                 59.9  15.3 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#>  3 A038423 D1                101.   14.1 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#>  4 A038423 D1                 99.0  13.2 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#>  5 A038423 D1                 99.0  12.3 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#>  6 A038423 D1                228.   12.7 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#>  7 A038423 D1                179.   12.0 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#>  8 A038423 D1                350.   13.0 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#>  9 A038423 D1                443.   11.8 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#> 10 A038423 D1                495.   11.3 Mn_Ca   mako shark  Isurus oxy…    30.1   -116. 2008-1…     157 M    
#> # … with 11,108 more rows, and abbreviated variable names ¹​scientific_name, ²​recapture_lat, ³​recapture_long,
#> #   ⁴​recapture_date, ⁵​recapture_fl

if (FALSE) {
### load package
library(dplyr)
library(ggplot2)

### vertebrae ba/ca
ggplot(data = dat[which(dat$element == "Ba_Ca"), ],
       aes(distance, value)) +
  geom_line(aes(colour = scientific_name, group = fish_id),
            show.legend = F, na.rm = T) +
  facet_grid(scientific_name ~ ., scales = "free_y") +
  xlab("Distance from core to edge of shark vertebrae (um)") +
  ylab("Ba/Ca (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### vertebrae mn/ca
ggplot(data = dat[which(dat$element == "Mn_Ca"), ],
       aes(distance, value)) +
  geom_line(aes(colour = scientific_name, group = fish_id),
            show.legend = F, na.rm = T) +
  facet_grid(scientific_name ~ ., scales = "free_y") +
  xlab("Distance from core to edge of shark vertebrae (um)") +
  ylab("Mn/Ca (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### vertebrae sr/ca
ggplot(data = dat[which(dat$element == "Sr_Ca"), ],
       aes(distance, value)) +
  geom_line(aes(colour = scientific_name, group = fish_id),
            show.legend = F, na.rm = T) +
  facet_grid(scientific_name ~ ., scales = "free_y") +
  xlab("Distance from core to edge of shark vertebrae (um)") +
  ylab("Sr/Ca (mmol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}