Otolith geochemistry reflects life histories of Pacific bluefin tuna

Otolith biominerals of large Pacific bluefin tuna Thunnus orientalis (PBT) collected from the western, eastern, and south Pacific Ocean, were analyzed for a suite of trace elements across calcified/proteinaceous growth zones to investigate patterns across ontogeny.

bt.mohan.2022

Format

The data frame 246,375 × 11 contains the following columns:

pbt_id	character	pacific bluefin tuna Thunnus orientalis identifier

distance	numeric	distance from otolith core to edge (um)

value	numeric	otolith element(E):Ca (Ba/Ca in umol/mol; Li/Ca in umol/mol; Mg/Ca in umol/mol; Mn/Ca in umol/mol; Sr/Ca in mmol/mol; Zn/Ca in umol/mol )

elements	character	bulk elements

date	character	date of capture

ocean	character	ocean of capture

region	character	region of capture

sex	character	fish sex

fl	numeric	fork length (cm)

age	character	estimated age (year)

gg_weight	integer	gilled and gutted (gg) weight (kg)

Details

Elemental profiles of individual Pacific bluefin tuna Thunnus orientalis (PBT) otoliths collected from the Eastern Pacific (EPO), Western Pacific (WPO) and South Pacific (SPO); different colored lines represent individual fish from each region.

Instrument: LA-ICP-MS (laser ablation-inductively coupled plasma mass spectrometry)

Beam diameter: 50um

Scan speed: 5um/s

Reference materials: NIST-612 (National Institutes of Standards and Technology glass standard) andMACS-3 (United States Geological Survey microanalytical carbonate standard pressed pellet)

Source

Mohan, J. A., Dewar, H., Snodgrass, O. E., Miller, N. R., Tanaka, Y., Ohshimo, S., ... & Wells, R. D. (2022). Otolith geochemistry reflects life histories of Pacific bluefin tuna. Plos one, 17(10), e0275899. https://doi.org/10.1371/journal.pone.0275899

Traversing the paper's information via Semantic Scholar ID 21574f9994add809b1d6c495a60d706ce070a552 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, bulk elements, Ba/Ca, Li/Ca, Mg/Ca, Mn/Ca, Sr/Ca, Zn/Ca

Examples

### copy data into 'dat'
dat <- bt.mohan.2022
tibble::tibble(dat)
#> # A tibble: 246,375 × 11
#>    pbt_id distance value elements     date ocean        region                    sex      fl   age gg_weight
#>    <chr>     <dbl> <dbl> <chr>       <int> <chr>        <chr>                     <chr> <int> <int>     <int>
#>  1 EP01       0     3.20 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#>  2 EP01       2.02  3.42 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#>  3 EP01       4.03  3.67 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#>  4 EP01       6.05  3.89 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#>  5 EP01       8.06  3.89 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#>  6 EP01      10.1   4.10 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#>  7 EP01      12.1   3.89 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#>  8 EP01      14.1   3.92 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#>  9 EP01      16.1   3.73 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#> 10 EP01      18.1   3.64 Li_Ca    20170920 East Pacific San Deigo, Califronia USA M       197     9       156
#> # … with 246,365 more rows

if (FALSE) {
### load package
library(dplyr)
library(ggplot2)

### Otolith ba/ca
ggplot(data = dat[which(dat$elements == "Ba_Ca"), ], aes(distance, value)) +
  geom_line(aes(colour = ocean, group = pbt_id), show.legend = F, na.rm = T) +
  facet_grid(ocean ~ ., scales = "free_y") +
  xlab("Distance from otolith core to edge (um)") +
  ylab("Otolith Ba/Ca (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Otolith li/ca
ggplot(data = dat[which(dat$elements == "Li_Ca"), ], aes(distance, value)) +
  geom_line(aes(colour = ocean, group = pbt_id), show.legend = F, na.rm = T) +
  facet_grid(ocean ~ ., scales = "free_y") +
  xlab("Distance from otolith core to edge (um)") +
  ylab("Otolith Li/Ca (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Otolith mg/ca
ggplot(data = dat[which(dat$elements == "Mg_Ca"), ], aes(distance, value)) +
  geom_line(aes(colour = ocean, group = pbt_id), show.legend = F, na.rm = T) +
  facet_grid(ocean ~ ., scales = "free_y") +
  xlab("Distance from otolith core to edge (um)") +
  ylab("Otolith Mg/Ca (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Otolith mn/ca
ggplot(data = dat[which(dat$elements == "Mn_Ca"), ], aes(distance, value)) +
  geom_line(aes(colour = ocean, group = pbt_id), show.legend = F, na.rm = T) +
  facet_grid(ocean ~ ., scales = "free_y") +
  xlab("Distance from otolith core to edge (um)") +
  ylab("Otolith Mn/Ca (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Otolith sr/ca
ggplot(data = dat[which(dat$elements == "Sr_Ca"), ], aes(distance, value)) +
  geom_line(aes(colour = ocean, group = pbt_id), show.legend = F, na.rm = T) +
  facet_grid(ocean ~ ., scales = "free_y") +
  xlab("Distance from otolith core to edge (um)") +
  ylab("Otolith Sr/Ca (mmol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Otolith zn/ca
ggplot(data = dat[which(dat$elements == "Zn_Ca"), ], aes(distance, value)) +
  geom_line(aes(colour = ocean, group = pbt_id), show.legend = F, na.rm = T) +
  facet_grid(ocean ~ ., scales = "free_y") +
  xlab("Distance from otolith core to edge (um)") +
  ylab("Otolith Zn/Ca (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}