Quantifyingmigratory capacity and dispersal of the invasive tench

The study of distribution and dispersal of invasive fishes is challenging during the early stages of invasion. Quantification of trace elements incorporated into fish hard parts represents an innovative technique for this task. Otolith chemistry has been used to describe fish stock structure, migratory behaviour and to support the management of several specie.

bt.morissette.2021

Format

The data frame 31,143 × 11 contains the following columns:

id	character	identifier of individual fish

site	character	site of capture

zone	character	data were separated in three distinct zones: core(C) of the otolith (first 10 analysis points from the core), otolith margin(E) (10 last analysis points) and complete transect(T).

distance_core	numeric	distance from the core of the otolith to edge(um)

srca	numeric	sr/ca (mmol/mol)

baca	numeric	ba/ca (mmol/mol)

mgca	numeric	mg/ca (mmol/mol)

mnca	numeric	mn/ca (mmol/mol)

kca	numeric	k/ca (mmol/mol)

feca	numeric	fe/ca (mmol/mol)

Details

The dataset contains core-to-edge transect of invasive tench (Tinca tinca) otolith.

Instrument: LA-ICP-MS (laser ablation-inductively coupled plasma mass spectrometry)

Beam diameter: 19um

Scan speed: 5um/s

Reference materials: NIST-610 (National Institutes of Standards and Technology glass standard) and MACS-3 (United States Geological Survey microanalytical carbonate standard pressed pellet)

Source

Morissette, O. Lecomte, F. Vachon, N. Drouin, A. & Sirois, P. (2021). Quantifying migratory capacity and dispersal of the invasive Tench (Tinca tinca) in the St. Lawrence River using otolith chemistry. Canadian Journal of Fisheries and Aquatic Sciences, 78(11), 1628-1638.https://doi.org/10.1139/cjfas-2020-0460

Data availability are available at https://doi.org/10.6084/m9.figshare.14406863.v1

Traversing the paper's information via Semantic Scholar ID 001378a49553ea5076562f9c41333cca74da56a8 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, bulk elements, Sr/ca, Ba/ca, Mg/ca, Mn/ca, Na/ca, K/ca, Fe/ca

Examples

### copy data into 'dat'
dat <- bt.morissette.2021
tibble::tibble(dat)
#> # A tibble: 31,143 × 11
#>    id      site           zone  distance_core  srca    baca  mgca  mnca  naca   kca  feca
#>    <chr>   <chr>          <chr>         <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 TITI-01 Haut-Richelieu C              0    0.561 0.00317  1.73 0.142  7.06 0.355 0.334
#>  2 TITI-01 Haut-Richelieu C              5.34 0.563 0.00355  1.65 0.149  6.81 0.336 0.351
#>  3 TITI-01 Haut-Richelieu C             10.7  0.562 0.00338  1.65 0.124  6.47 0.340 0.324
#>  4 TITI-01 Haut-Richelieu C             16.0  0.511 0.00367  1.54 0.130  5.90 0.347 0.329
#>  5 TITI-01 Haut-Richelieu C             21.4  0.545 0.00344  1.71 0.148  6.35 0.335 0.301
#>  6 TITI-01 Haut-Richelieu C             26.7  0.532 0.00398  1.63 0.136  6.93 0.338 0.314
#>  7 TITI-01 Haut-Richelieu C             32.0  0.499 0.00353  1.67 0.138  6.71 0.332 0.326
#>  8 TITI-01 Haut-Richelieu C             37.4  0.556 0.00390  1.56 0.143  6.18 0.332 0.298
#>  9 TITI-01 Haut-Richelieu C             42.7  0.572 0.00474  1.73 0.150  6.94 0.347 0.339
#> 10 TITI-01 Haut-Richelieu C             48.1  0.489 0.00385  1.75 0.140  6.66 0.334 0.310
#> # … with 31,133 more rows

if (FALSE) {
### loading packages
library(dplyr)
library(ggplot2)

### Taking Sr/Ca of otolith
### Otolith core
ggplot(data = dat[which(dat$zone == "C"),],aes(site, srca))+
  geom_boxplot()+
  ylab("Sr/Ca (mmol/mol)")+
  xlab("Site of capture")+
  labs(title = "Otolith core")

### Otolith edge
ggplot(data = dat[which(dat$zone == "E"),],aes(site, srca))+
  geom_boxplot()+
  ylab("Sr/Ca (mmol/mol)")+
  xlab("Site of capture")+
  labs(title = "Otolith edge")

### Core-to-edge transect
ggplot(data = dat[which(dat$zone == "T"),],aes(distance_core, srca))+
  geom_line(aes(col = site, group = id),show.legend = F,linewidth = 0.02, na.rm = T)+
  facet_grid(site~.)+
  labs(
    x = expression(paste("Distance from the core of the otolith (", mu, "m)", sep = "")),
    y = "Sr/Ca (mmol/mol)"
  )+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}