Studies of identifying the geographical origin of fish

In this study, the authors used Sr isotopic markers recorded in otoliths of wild and farmed commercialized pirarucu (Arapaima spp.) to evaluate their prediction potential to trace the fishes origin.

bt.pereira.2019

Format

The data frame 4,148 × 5 contains the following columns:

specimen_code	character	specimen id

distance	integer	distance from fish otolith cores – edge transects (unit: micron)

oto_sr8786	numeric	variation of 87Sr/86Sr measured by LAfs-MC-ICPMS

fish_origin	character	wild fish and farmed fish

region	character	region of captured

Details

The dataset contains wild and farmed commercialized pirarucu (Arapaima spp.) otolith strontium isotope profiles. Variation of 87Sr/86Sr values measured by LAfs-MC-ICPMS on wild fish otolith cores – edge transects.

Wild fish specimens that lived in environments with the largest fluctuation of river water Sr isotope ratios over time presented the largest Sr isotope variations in otoliths.

Source

Pereira, L. A., Santos, R. V., Hauser, M., Duponchelle, F., Carvajal, F., Pecheyran, C., ... & Pouilly, M. (2019). Commercial traceability of Arapaima spp. fisheries in the Amazon basin: can biogeochemical tags be useful?. Biogeosciences, 16(8), 1781--1797. https://doi.org/10.5194/bg-16-1-2019

Data availability are available at https://doi.org/10.5281/zenodo.2646436

Traversing the paper's information via Semantic Scholar ID 313992f560c0c54460e235a4963016148dd61033 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, isotopic ratios, Sr8786

Examples

### copy data into 'dat'
dat <- bt.pereira.2019
tibble::tibble(dat)
#> # A tibble: 4,148 × 5
#>    specimen_code distance oto_sr8786 fish_origin region        
#>    <chr>            <int>      <dbl> <chr>       <chr>         
#>  1 I1                   0      0.710 wild        Central Amazon
#>  2 I1                   5      0.710 wild        Central Amazon
#>  3 I1                  10      0.710 wild        Central Amazon
#>  4 I1                  15      0.710 wild        Central Amazon
#>  5 I1                  20      0.711 wild        Central Amazon
#>  6 I1                  25      0.710 wild        Central Amazon
#>  7 I1                  30      0.710 wild        Central Amazon
#>  8 I1                  35      0.710 wild        Central Amazon
#>  9 I1                  40      0.710 wild        Central Amazon
#> 10 I1                  45      0.710 wild        Central Amazon
#> # … with 4,138 more rows

if (FALSE) {
### Sr profile figure
library(dplyr)
library(ggplot2)

ggplot(data = dat,aes(distance, oto_sr8786))+
  geom_line(aes(col = fish_origin),show.legend = F,linewidth = 0.02, na.rm = T)+
  facet_grid(region~ fish_origin)+
  labs(
    x = (expression(
      paste("distance from fish otolith cores – edge transects (", mu, "m)",
            sep = ""))),
    y = (expression(paste(
      {}^"87",
      "Sr/",
      {}^"86",
      "Sr"
    )))
  )+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

}