Eye-lens isotopes of teleost and elasmobranch

It is shown that fish eye lens protein is an attractive host for recovery of stable isotope life histories, particularly for juvenile life stages, and especially in elasmobranchs lacking otoliths, but interpretation of lens-based records is complicated by species-specific uncertainties associated with lens growth rates.

bt.quaeckdavies.2018

Format

The list contains 5 data frame the following columns:

species	character	teleost and elasmobranch studied

tl	numeric	body length (mm)

lens_d	numeric	lens diameter (mm)

d13c	numeric	stable isotope d13C values

d13cc	numeric	muscle d13C values have been corrected for lipid contents

d15n	numeric	stable isotope d15N values

tissue	character	muscle (M) and lens (L)

fish_id	character	individual fish

radius1	numeric	maximum radial of the subsampled lens

radius2	numeric	minimum radial of the subsampled lens

diam1	numeric	maximum diamater measurements of the subsampled lens

diam2	numeric	minimum diamater measurements of the subsampled lens

recovered_length	numeric	estimated fish length at a given lens diameter

Details

This dataset contains 5 data frame:

TL_LD: Paired eye lens diameter and total body length data. Paired lens diameter (Lens D, mm) and body length (TL, mm) measurements for four species, Squalus acanthias, Coryphaenoides rupestris, Lamna nasus and Aphanopus carbo.

Teleost_lensCore: Stable isotope values of lens cores in three species of fishes. Carbon (d13C) and nitrogen (d15N) values for the central regions of eye lenses in Coryphaenoides rupesitris (grenadier), Lamna nasus (porbeagle) and Aphanopus carbo (scabbard).

CrupL_M: Eye lens and muscle isotope data for juvenile C. rupestris. Eye lens and muscle isotope data (d13C and d15N values) for juvenile C. rupestris from the Rockall Trough.

Acarbo_lens: Isotopic transects across lenses of A. carbo. Stable isotope compositions (d13C d15N values) of sequential lens protein samples in 7 Aphanopus carbo individuals captured in long-line fisheries in west Portugal. Fish ID indicates the individual fish, radius 1 and 2 and diameter 1 and 2 indicate the maximum and minimum radial and diamater measurements of the subsampled lens respectively. Reconstructed length is the estimated fish length at a given lens diameter estimated from allometic scaling relationships given in the text.

Embryo: Stable isotope composition of paired embryonic muscle and lens tissues. Stable isotope (d13C and d15N) values for muscle (M) and lens (L) proteins in 19 embryonic Squalus acanthias. d13Cc values indicate that muscle d13C values have been corrected for lipid contents following Reum (2011).

Instrument: IRMS (isotope ratio mass spectrometer)

Source

Quaeck-Davies, K. Bendall, V. A. MacKenzie, K. M. Hetherington, S. Newton, J. & Trueman, C. N. (2018). Teleost and elasmobranch eye lenses as a target for life-history stable isotope analyses. PeerJ, 6, e4883. https://doi.org/10.7717/peerj.4883

Traversing the paper's information via Semantic Scholar ID aa714946f791e2e9b834cf8a0b94674a98665105 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, bulk elements, Sr/Ca

Examples

### copy data into 'dat'
dat <- bt.quaeckdavies.2018

### TL_LD data frame
tibble::tibble(dat$TL_LD)
#> # A tibble: 204 × 3
#>    species              tl lens_d
#>    <chr>             <dbl>  <dbl>
#>  1 Squalus acanthias   870   9   
#>  2 Squalus acanthias  1100  13   
#>  3 Squalus acanthias   770   9.5 
#>  4 Squalus acanthias   840  10   
#>  5 Squalus acanthias   650   8   
#>  6 Squalus acanthias  1070  11.8 
#>  7 Squalus acanthias   780  10   
#>  8 Squalus acanthias   830  10.2 
#>  9 Squalus acanthias   400   4.75
#> 10 Squalus acanthias   530   7.5 
#> # … with 194 more rows

### Teleost_lensCore data frame
tibble::tibble(dat$Teleost_lensCore)
#> # A tibble: 29 × 3
#>    species                    d13c  d15n
#>    <chr>                     <dbl> <dbl>
#>  1 Coryphaenoides rupesitris -20.6  4.72
#>  2 Coryphaenoides rupesitris -17.9  6.74
#>  3 Coryphaenoides rupesitris -19.9  6.06
#>  4 Coryphaenoides rupesitris -19.5  7.46
#>  5 Coryphaenoides rupesitris -19.7  6.49
#>  6 Coryphaenoides rupesitris -18.3  4.62
#>  7 Coryphaenoides rupesitris -19.5  7.97
#>  8 Coryphaenoides rupesitris -19.6  6.74
#>  9 Coryphaenoides rupesitris -19.1  5.89
#> 10 Coryphaenoides rupesitris -20.5  6.36
#> # … with 19 more rows

### CrupL_M data frame
tibble::tibble(dat$CrupL_M)
#> # A tibble: 22 × 4
#>    tissue  d13c  d15n species                  
#>    <chr>  <dbl> <dbl> <chr>                    
#>  1 Lens   -20.6  4.93 Coryphaenoides rupesitris
#>  2 Lens   -17.9  6.63 Coryphaenoides rupesitris
#>  3 Lens   -19.9  6.20 Coryphaenoides rupesitris
#>  4 Lens   -19.5  7.55 Coryphaenoides rupesitris
#>  5 Lens   -19.7  6.60 Coryphaenoides rupesitris
#>  6 Lens   -18.3  4.57 Coryphaenoides rupesitris
#>  7 Lens   -19.5  8.05 Coryphaenoides rupesitris
#>  8 Lens   -19.6  6.84 Coryphaenoides rupesitris
#>  9 Lens   -19.1  5.93 Coryphaenoides rupesitris
#> 10 Lens   -20.5  6.57 Coryphaenoides rupesitris
#> # … with 12 more rows

### Acarbo_lens data frame
tibble::tibble(dat$Acarbo_lens)
#> # A tibble: 73 × 10
#>    fish_id sample  d13c  d15n radius1 radius2 diam1 diam2 recovered_length species        
#>    <chr>   <chr>  <dbl> <dbl>   <dbl>   <dbl> <dbl> <dbl>            <dbl> <chr>          
#>  1 C512    C512B  -18.4  15.2    -3.5     4.1   7     8.2             612. Aphanopus carbo
#>  2 C512    C512C  -18.7  14.6    -2.8     3.4   5.6   6.8             536. Aphanopus carbo
#>  3 C512    C512D  -19.0  14.2    -2.1     2.7   4.2   5.4             460. Aphanopus carbo
#>  4 C512    C512E  -19.0  13.5    -1.4     2     2.8   4               384. Aphanopus carbo
#>  5 C512    C512F  -19.4  12.2    -0.7     1.3   1.4   2.6             309. Aphanopus carbo
#>  6 C512    C512G  -19.8  10.5     0       0.6   0     1.2             233. Aphanopus carbo
#>  7 C512    C512H  -19.4  12.6     0.7     1.3   1.4   2.6             309. Aphanopus carbo
#>  8 C512    C512I  -18.7  13.4     1.4     2     2.8   4               384. Aphanopus carbo
#>  9 C512    C512J  -19.0  14.0     2.1     2.7   4.2   5.4             460. Aphanopus carbo
#> 10 C405    C405A  -17.9  15.2    -3.5     4.1   7     8.2             612. Aphanopus carbo
#> # … with 63 more rows

### Embryo data frame
tibble::tibble(dat$Embryo)
#> # A tibble: 38 × 6
#>       id tissue  d15n d13cc d13c2 species          
#>    <int> <chr>  <dbl> <dbl> <dbl> <chr>            
#>  1    11 M       12.7 -15.7 -16.2 Squalus acanthias
#>  2    16 M       11.2 -16.0 -16.6 Squalus acanthias
#>  3    28 M       13.8 -15.8 -16.4 Squalus acanthias
#>  4    32 M       11.6 -16.5 -17.0 Squalus acanthias
#>  5    36 M       14.3 -15.4 -15.9 Squalus acanthias
#>  6    37 M       12.0 -16.3 -16.8 Squalus acanthias
#>  7    38 M       11.7 -16.0 -16.4 Squalus acanthias
#>  8    39 M       12.5 -15.5 -16.1 Squalus acanthias
#>  9    54 M       11.7 -15.8 -16.3 Squalus acanthias
#> 10    56 M       12   -16.1 -16.6 Squalus acanthias
#> # … with 28 more rows

if (FALSE) {
### load package
library(dplyr)
library(ggplot2)

### TL_LD data frame
ggplot(data = dat$TL_LD, aes(tl, lens_d)) +
  geom_point(aes(colour = species), na.rm = T, size = 2) +
  geom_smooth(aes(colour = species),
    show.legend = F, formula = "y ~ x",
    method = "lm", na.rm = T
  ) +
  xlab("Total Length (mm)") +
  ylab("Lens Diameter (mm)") +
  theme_bw() +
  theme(
    legend.position = "top",
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Teleost_lensCore data frame
ggplot(data = dat$Teleost_lensCore, aes(d13c, d15n)) +
  geom_point(aes(colour = species), na.rm = T, size = 3) +
  xlab(expression(paste("Lens cores ", delta^{
    13
  }, "C"))) +
  ylab(expression(paste("Lens cores ", delta^{
    15
  }, "N"))) +
  theme_bw() +
  theme(
    legend.position = "top",
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### CrupL_M data frame
ggplot(data = dat$CrupL_M, aes(tissue, d13c)) +
  geom_boxplot(aes(colour = tissue), na.rm = T, show.legend = F) +
  geom_point(aes(colour = tissue), na.rm = T, show.legend = F) +
  xlab("Tissue") +
  ylab(expression(paste(delta^{
    13
  }, "C"))) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

ggplot(data = dat$CrupL_M, aes(tissue, d15n)) +
  geom_boxplot(aes(colour = tissue), na.rm = T, show.legend = F) +
  geom_point(aes(colour = tissue), na.rm = T, show.legend = F) +
  xlab("Tissue") +
  ylab(expression(paste(delta^{
    15
  }, "N"))) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Acarbo_lens data frame
ggplot(data = dat$Acarbo_lens, aes(radius1, d13c)) +
  geom_point(aes(group = fish_id, colour = fish_id), na.rm = T, show.legend = F) +
  geom_line(aes(group = fish_id, colour = fish_id), na.rm = T, show.legend = F) +
  xlab("Lens radius (mm)") +
  ylab(expression(paste(delta^{
    13
  }, "C"))) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

ggplot(data = dat$Acarbo_lens, aes(radius1, d15n)) +
  geom_point(aes(group = fish_id, colour = fish_id), na.rm = T, show.legend = F) +
  geom_line(aes(group = fish_id, colour = fish_id), na.rm = T, show.legend = F) +
  xlab("Lens radius (mm)") +
  ylab(expression(paste(delta^{
    15
  }, "N"))) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### Embryo data frame
ggplot(data = dat$Embryo, aes(tissue, d13cc)) +
  geom_boxplot(aes(colour = tissue), na.rm = T, show.legend = F) +
  geom_point(aes(colour = tissue), na.rm = T, show.legend = F) +
  xlab("Tissue") +
  ylab(expression(paste(delta^{
    13
  }, "C"))) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

ggplot(data = dat$Embryo, aes(tissue, d15n)) +
  geom_boxplot(aes(colour = tissue), na.rm = T, show.legend = F) +
  geom_point(aes(colour = tissue), na.rm = T, show.legend = F) +
  xlab("Tissue") +
  ylab(expression(paste(delta^{
    15
  }, "N"))) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}