Life history responses to climate variability in sardine populations

In this study, the authors investigated the early life history traits and their responses to seawater temperature variation of sardines in western and eastern boundary current systems to better understand the factors influencing the populations’ apparent opposite responses to decadal scale anomalies in ocean temperature.

bt.sakamoto.2022

Format

The data frame 1,558 × 7 contains the following columns:

region_id	character	"JP" is JP sardine and "CA is CA sardine

fish_id	character	fish identifier

year_class	integer	year of hach

station_id	character	number of station in cuise surveys or ID of landing port

age_range	character	specifies the age range that was micromilled that are differently defined for JP and CA sardine. For JP: A: 0-30 dph, B: 31-45 dph, C: 46-60 dph, D: 61-75 dph, E: 76-90 dph, F: 91-105 dph, G: 106-120 dph and for CA: A: 0-30 dph, B: 31-60 dph, C: 61-90 dph, D: 91-120 dph, E: 121-150 dph

d13c	numeric	stable carbon isotope ratio relative to VPDB in permil

d18o	numeric	stable oxygen isotope ratio relative to VPDB in permil

Details

Stable carbon and oxygen isotope ratios of powders extracted from otoliths of JP and CA sardines by micromilling them. The d13C and d18O values were reported in delta-notation relative to the Vienna Pee Dee Belemnite (VPDB) scale and are given as a permil value. Analytical precisions were better than 0.10 permil and 0.17 permil for d18O, and better than 0.10 permil and 0.15 permil for d13C, respectively. The acid fractionation factor of calcite was used. Because the difference between the acid fractionation factor of calcite and aragonite is temperature dependent, 0.09 permil was subtracted from the d18O value determined by DELTA V to adjust for the different response temperatures.

Instrument: IRMS (isotope ratio mass spectrometer)

Source

Sakamoto, T., Takahashi, M., Chung, M. T., Rykaczewski, R. R., Komatsu, K., Shirai, K., ... & Higuchi, T. (2022). Contrasting life-history responses to climate variability in eastern and western North Pacific sardine populations. Nature Communications, 13(1), 5298. https://doi.org/10.1038/s41467-022-33019-z

Data availability are available at https://doi.org/10.5281/zenodo.6983520

Traversing the paper's information via Semantic Scholar ID a30c43db35f53fcbb1f5c364cd89ad6f7eb9afb5 using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, stable isotope, d13C, d18O

Examples

### copy data into 'dat'
dat <- bt.sakamoto.2022
tibble::tibble(dat)
#> # A tibble: 1,558 × 7
#>    region_id fish_id          year_class station_id age_range  d13c    d18o
#>    <chr>     <chr>                 <int> <chr>      <chr>     <dbl>   <dbl>
#>  1 JP        2015_9_HH_T9_1R        2015 2015_T9    B         -6.93 -0.240 
#>  2 JP        2015_9_HH_T9_1R        2015 2015_T9    C         -7.90 -0.148 
#>  3 JP        2015_9_HH_T9_1R        2015 2015_T9    D         -7.77 -0.0269
#>  4 JP        2015_9_HH_T9_1R        2015 2015_T9    E         -6.47 -0.113 
#>  5 JP        2015_9_HH_T9_1R        2015 2015_T9    F         -5.76  0.154 
#>  6 JP        2015_9_HH_T9_1R        2015 2015_T9    G         -5.31  0.294 
#>  7 JP        2015_9_HH_T9_10L       2015 2015_T9    A         -7.60 -0.266 
#>  8 JP        2015_9_HH_T9_10L       2015 2015_T9    B         -8.57 -0.201 
#>  9 JP        2015_9_HH_T9_10L       2015 2015_T9    C         -8.86 -0.172 
#> 10 JP        2015_9_HH_T9_10L       2015 2015_T9    D         -7.32  0.123 
#> # ℹ 1,548 more rows

if (FALSE) {
### loading packages
library(dplyr)
library(ggplot2)

### d13C profile figure
ggplot(data = dat,aes(age_range, d13c))+
  geom_boxplot(aes(col = region_id),show.legend = F,size = 0.02, na.rm = T)+
  facet_grid(.~ region_id)+
  labs(
    x = ("Age range"),
    y = (expression(delta * ""^13 * "C" * " (‰)"))
  )+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### d18O profile figure
ggplot(data = dat,aes(age_range, d18o))+
  geom_boxplot(aes(col = region_id),show.legend = F,size = 0.02, na.rm = T)+
  facet_grid(.~ region_id)+
  labs(
    x = ("Age range"),
    y = (expression(delta * ""^18 * "O" * " (‰)"))
  )+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}