Shark tooth collagen stable isotopes as ecological proxies

Within-individual variability of tooth stable isotope values suggests teeth of different ages may serve as ecological chronologies, which could be applied to studies on migration and individual-level diet variation across diverse time scales.

bt.shipley.2021

Format

The data frame 33 × 5 contains the following columns:

fish_id	character	shark identifier

jaw_position	character	jaw position

file_position	numeric	location of the row relative to the centre, lower numbers represent older, more mature teeth and higher numbers represent younger teeth

d15n	numeric	carbon isotope ratio of shark tooth collagen

d13c	numeric	nitrogen isotope ratio of shark tooth collagen

Details

This dataset contains measured isotope values (expressed in ‰) for shark (shortfin mako shark Isurus oxyrinchus and sandbar shark Carcharhinus plumbeus) tooth collagen, for individual teeth across different rows and tooth positions.

Instrument: EA-IRMS (elemental analyzer–isotope ratio mass spectrometry)

Source

Shipley, O. N., Henkes, G. A., Gelsleichter, J., Morgan, C. R., Schneider, E. V., Talwar, B. S., & Frisk, M. G. (2021). Shark tooth collagen stable isotopes (δ15N and δ13C) as ecological proxies. Journal of Animal Ecology, 90(9), 2188--2201. https://doi.org/10.1111/1365-2656.13518

Data availability are available at https://doi.org/10.5061/dryad.j6q573ncs

Traversing the paper's information via Semantic Scholar ID 4e0d4172c63f18f7ae22c4800031c4e96db1a2cc using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

tooth, stable isotope, d13c, d15n

Examples

### copy data into 'dat'
dat <- bt.shipley.2021
tibble::tibble(dat)
#> # A tibble: 33 × 5
#>    fish_id             jaw_position file_position  d15n  d13c
#>    <chr>               <chr>                <int> <dbl> <dbl>
#>  1 Shortfin mako shark Bottom Jaw               1  12.8 -13.4
#>  2 Shortfin mako shark Bottom Jaw               3  12.5 -13.4
#>  3 Shortfin mako shark Bottom Jaw               5  11.9 -13.3
#>  4 Shortfin mako shark Bottom Jaw               7  11.1 -13.4
#>  5 Shortfin mako shark Top Jaw                  1  13.4 -13.8
#>  6 Shortfin mako shark Top Jaw                  3  12.8 -13.6
#>  7 Shortfin mako shark Top Jaw                  5  13.5 -14.1
#>  8 Shortfin mako shark Top Jaw                  7  12.2 -13.3
#>  9 Shortfin mako shark Top Jaw                  9  12.1 -13.6
#> 10 Sandbar Shark #1    Bottom Jaw               1  15.4 -14.1
#> # ℹ 23 more rows

if (FALSE) {
### loading packages
library(dplyr)
library(ggplot2)

### d13C profile figure
ggplot(data = dat, aes(file_position, d13c)) +
  geom_line(aes(col = fish_id), show.legend = F, linewidth = 0.02, na.rm = T) +
  geom_point(aes(col = fish_id), show.legend = F, na.rm = T) +
  facet_wrap(. ~ jaw_position) +
  labs(
    x = ("Location of the row relative to the centre"),
    y = (expression(delta * ""^13 * "C" * " (‰)"))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### d15N profile figure
ggplot(data = dat, aes(file_position, d15n)) +
  geom_line(aes(col = fish_id), show.legend = F, linewidth = 0.02, na.rm = T) +
  geom_point(aes(col = fish_id), show.legend = F, na.rm = T) +
  facet_wrap(. ~ jaw_position) +
  labs(
    x = ("Location of the row relative to the centre"),
    y = (expression(delta * ""^15 * "N" * " (‰)"))
  ) +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}