Reconstructing the migratory behavior of juvenile chinook salmon

The authors used otolith 87Sr/86Sr in adult chinook salmon (Oncorhynchus tshawytcha) returning to the Stanislaus River in the California Central Valley (USA) to reconstruct the sizes at which they outmigrated as juveniles in a wetter (2000) and drier (2003) year.

bt.sturrock.2015

Format

The data frame 3,654 × 11 contains the following columns:

sample_id	character	sample identifier

collection_date	character	sample collection date

length	integer	length at collection (cm)

age	integer	age at collection (years)

sex	integer	sex (female or male)

outmigration_year	integer	outmigration year

distance	numeric	distance from otolith core (um)

oto_sr8786	numeric	otolith 87Sr/86Sr value (spot analysis)

oto_sr8786_2se	numeric	instrument error

oto_srV	numeric	strontium signal

laser_properties	character	laser properties (spot size, power, repetition rate, dwell time, spacing)

Details

The dataset contains 87Sr/86Sr values measured in the otoliths of adult chinook salmon sampled in the Stanislaus River, California, between 2001 and 2006. Fish age was estimated using scale annuli, and individuals that outmigrated in 2000 and 2003 were retained for the current analysis (see "outmigration year"). Each otolith contains multiple 87Sr/86Sr values, representing spot analyses sampled along a standardized 90° transect from the otolith core towards the otolith edge.

Instrument: LA-MC-ICP-MS (laser ablation multi-collector inductively coupled plasma mass spectrometry)

Source

Sturrock, A. M., Wikert, J. D., Heyne, T., Mesick, C., Hubbard, A. E., Hinkelman, T. M., ... & Johnson, R. C. (2015). Reconstructing the migratory behavior and long-term survivorship of juvenile Chinook salmon under contrasting hydrologic regimes. PLoS One, 10(5), e0122380. https://doi.org/10.1371/journal.pone.0122380

Data availability are available at https://doi.org/10.5061/dryad.c56rk

Traversing the paper's information via Semantic Scholar ID bd4c80e8eef6f81369f81af5c12016f30ffde72e using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, stable isotope, Sr8786

Examples

### copy data into 'dat'
dat <- bt.sturrock.2015
tibble::tibble(dat)
#> # A tibble: 3,654 × 11
#>    sample_id collection_date length   age sex   outmigration_year distance oto_sr8786 oto_sr8786_2se
#>    <chr>     <chr>            <int> <int> <chr>             <int>    <dbl>      <dbl>          <dbl>
#>  1 SR11234   2001-11-30          58     2 F                  2000     30.7      0.708       0.0003  
#>  2 SR11234   2001-11-30          58     2 F                  2000     75.3      0.708       0.000178
#>  3 SR11234   2001-11-30          58     2 F                  2000    125.       0.708       0.000108
#>  4 SR11234   2001-11-30          58     2 F                  2000    177.       0.707       0.000088
#>  5 SR11234   2001-11-30          58     2 F                  2000    229.       0.707       0.000116
#>  6 SR11234   2001-11-30          58     2 F                  2000    276.       0.707       0.00007 
#>  7 SR11234   2001-11-30          58     2 F                  2000    324.       0.707       0.000078
#>  8 SR11234   2001-11-30          58     2 F                  2000    376.       0.708       0.000188
#>  9 SR11234   2001-11-30          58     2 F                  2000    423.       0.708       0.000196
#> 10 SR11234   2001-11-30          58     2 F                  2000    470.       0.709       0.00017 
#> # ℹ 3,644 more rows
#> # ℹ 2 more variables: oto_sr_v <dbl>, laser_properties <chr>

if (FALSE) {
### loading packages
library(dplyr)
library(ggplot2)

### Sr8786 profile figure
ggplot(data = dat,aes(distance, oto_sr8786))+
  geom_line(aes(col = sex),show.legend = F,linewidth = 0.02, na.rm = T)+
  facet_grid(.~ sex)+
  labs(
    x = expression(paste("Distance from core (", mu, "m)", sep = "")),
    y = expression(paste(
      {}^"87",
      "Sr/",
      {}^"86",
      "Sr"
    ))
  )+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}