Unnatural selection of salmon life histories in a modified riverscape

The expression and ultimate success of diverse salmon emigration behaviors in an anthropogenically altered California river system is quantified and management actions favoring any single phenotype could have negative evolutionary and demographic consequences, potentially reducing adaptability and population stability.

bt.sturrock.2020

Format

The data frame 16,463 × 12 contains the following columns:

Sample_ID	character	sample identifier

ASN	integer	numbers represent the agency ID given to each fish

Natal_region	character	natal region

Otolith_radius_um	numeric	otolith radius (um)

Otolith_87Sr86Sr	numeric	otolith strontium isotope (Sr87/Sr86) values

SE2	numeric	otolith strontium isotope (Sr87/Sr86) standard error

SrV	numeric	SrV

Outmigration_year	integer	year of Out-migration

Capture_FL_cm	numeric	fish fork length of captured

Capture_Age	integer	fish age of captured

Sex	character	fish sex(male or female)

CWT_no	integer	fish analyzed blind to validate natal assignment accuracy

Details

The dataset contains natal otolith Sr87/Sr86 values was used to predict provenance of the spawning adults using the methods described in (Barnett-Johnson et al. 2008, Sturrock et al. 2015).

Source

Rachel, B. J., Pearson, T. E., Ramos, F. C., Grimes, C. B., & Bruce MacFarlane, R. (2008). Tracking natal origins of salmon using isotopes, otoliths, and landscape geology. Limnology and Oceanography, 53(4), 1633-1642.

Sturrock, A. M., Carlson, S. M., Wikert, J. D., Heyne, T., Nusslé, S., Merz, J. E., ... & Johnson, R. C. (2020). Unnatural selection of salmon life histories in a modified riverscape. Global Change Biology, 26(3), 1235-1247. https://doi.org/10.1111/gcb.14896

Sturrock, A. M., Wikert, J. D., Heyne, T., Mesick, C., Hubbard, A. E., Hinkelman, T. M., ... & Johnson, R. C. (2015). Reconstructing the migratory behavior and long-term survivorship of juvenile Chinook salmon under contrasting hydrologic regimes. PLoS One, 10(5), e0122380.

Data availability are available at https://doi.org/10.5061/dryad.73n5tb2ss

Traversing the paper's information via Semantic Scholar ID af6fc0e877a33843d232372c7f038542e5621a1d using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, isotopic ratios, Sr8786

Examples

### copy data into 'dat'
dat <- bt.sturrock.2020
tibble::tibble(dat)
#> # A tibble: 16,463 × 12
#>    Sample_ID   ASN Natal_region Otolith_r…¹ Otoli…²     SE2   SrV Outmi…³ Captu…⁴ Captu…⁵ Sex   CWT_no
#>    <chr>     <int> <chr>              <dbl>   <dbl>   <dbl> <dbl>   <int>   <int>   <int> <chr>  <int>
#>  1 SR11234   11234 ""                  30.7   0.708 3   e-4  3.45    2000      58       2 F         NA
#>  2 SR11234   11234 ""                  75.3   0.708 1.78e-4  3.03    2000      58       2 F         NA
#>  3 SR11234   11234 ""                 125.    0.708 1.08e-4  3.30    2000      58       2 F         NA
#>  4 SR11234   11234 ""                 177.    0.707 8.8 e-5  3.98    2000      58       2 F         NA
#>  5 SR11234   11234 ""                 229.    0.707 1.16e-4  3.37    2000      58       2 F         NA
#>  6 SR11234   11234 ""                 276.    0.707 7   e-5  3.79    2000      58       2 F         NA
#>  7 SR11234   11234 "y"                324.    0.707 7.8 e-5  3.87    2000      58       2 F         NA
#>  8 SR11234   11234 "y"                376.    0.708 1.88e-4  4.01    2000      58       2 F         NA
#>  9 SR11234   11234 "y"                423.    0.708 1.96e-4  3.84    2000      58       2 F         NA
#> 10 SR11234   11234 ""                 470.    0.709 1.7 e-4  3.78    2000      58       2 F         NA
#> # … with 16,453 more rows, and abbreviated variable names ¹​Otolith_radius_um, ²​Otolith_87Sr86Sr,
#> #   ³​Outmigration_year, ⁴​Capture_FL_cm, ⁵​Capture_Age

if (FALSE) {
### loading packages
library(dplyr)
library(ggplot2)

### Sr8786 profile figure
ggplot(data = dat,aes(Otolith_radius_um, Otolith_87Sr86Sr))+
  geom_line(aes(col = Sex),show.legend = F,linewidth = 0.02, na.rm = T)+
  facet_grid(.~ Sex)+
  labs(
    x = (expression(paste("Otolith radius (", mu, "m)", sep = ""))),
    y = (expression(paste(
      {}^"87",
      "Sr/",
      {}^"86",
      "Sr"
    )))
  )+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}