Life history strategies of Mekong pangasiid catfishes*

In order to examine the life-history strategies of key pangasiid catfishes in the Mekong River, the authors collected otoliths from 126 specimens belonging to four species in the Lower Mekong Basin. The authors analysed the otolith microchemical structure using two methods (Laser Ablation-Inductively Coupled Plasma Mass Spectrometry and Scanning X-ray Fluorescence Microscopy) to identify evidence of diadromy (movements between the ocean and upstream waters).

bt.vu.2022a

Format

The data frame 27,522 × 13 contains the following columns:

fish_id	character	shark identifier

species	character	mekong pangasiid catfishes

distance	numeric	distance from core to edge of otolith (um)

value	numeric	otolith element(E):Ca (Ba/Ca x1000;Sr/Ca x1000)

element	character	sr_ca or ba_ca

technology	character	analysis technology

sampling_date	character	date of sampling

latitude	numeric	latitude of capture

longitude	numeric	longitude of capture

length	numeric	body length (cm)

weight	numeric	body weight (g)

sex	character	fish sex

sectioning_plane	character	sectioning plane

Details

The dataset contains element(E):Ca from otolith core–edge transects of Mekong pangasiid catfishes (Pangasius bocourti, Pangasius elongatus, Pangasius krempfi, and Pangasius mekongensis).

Instrument: LA-ICP-MS (laser ablation-inductively coupled plasma mass spectrometry);SXFM (scanning x-ray fluorescence imaging)

Beam diameter: 25-30um

Scan speed: 3-10um/s

Reference materials: NIST-612 (National Institutes of Standards and Technology glass standard) and MACS-3 (United States Geological Survey microanalytical carbonate standard pressed pellet)

Note: data extraction via WebPlotDigitizer tool (Title with * label).

Source

Vu, A. V., Baumgartner, L. J., Limburg, K. E., Doran, G. S., Mallen-Cooper, M., Gillanders, B. M., ... & Cowx, I. G. (2022). Life history strategies of Mekong pangasiid catfishes revealed by otolith microchemistry. Fisheries Research, 249, 106239.https://doi.org/10.1016/j.fishres.2022.106239

Vu, A. V. (2022). Diversity in migration patterns of selected Mekong River fishes. Charles Sturt University. https://researchoutput.csu.edu.au/en/publications/diversity-in-migration-patterns-of-selected-mekong-river-fishes

Traversing the paper's information via Semantic Scholar ID 8adc040f8c2d7913571ecf8f5cd2f3ae282c339f using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

otolith, bulk elements, Ba/Ca, Sr/Ca

Examples

### copy data into 'dat'
dat <- bt.vu.2022a
tibble::tibble(dat)
#> # A tibble: 27,522 × 13
#>    fish_id  species          dista…¹ value element techn…² sampl…³ longi…⁴ latit…⁵ length weight sex   secti…⁶
#>    <chr>    <chr>              <dbl> <dbl> <chr>   <chr>   <chr>     <dbl>   <dbl> <chr>   <dbl> <chr> <chr>  
#>  1 103-CLD2 Pangasius bocou…    1.71  2.81 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#>  2 103-CLD2 Pangasius bocou…    1.80  2.68 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#>  3 103-CLD2 Pangasius bocou…    7.86  2.56 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#>  4 103-CLD2 Pangasius bocou…    9.63  2.46 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#>  5 103-CLD2 Pangasius bocou…   14.5   2.34 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#>  6 103-CLD2 Pangasius bocou…   15.6   2.21 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#>  7 103-CLD2 Pangasius bocou…   23.6   2.05 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#>  8 103-CLD2 Pangasius bocou…   34.4   1.92 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#>  9 103-CLD2 Pangasius bocou…   42.6   1.83 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#> 10 103-CLD2 Pangasius bocou…   53.7   1.74 Sr_Ca   LA-ICP… 2018-0…    106.    9.51 34        600 Unkn… Frontal
#> # … with 27,512 more rows, and abbreviated variable names ¹​distance, ²​technology, ³​sampling_date, ⁴​longitude,
#> #   ⁵​latitude, ⁶​sectioning_plane

if (FALSE) {
### load package
library(dplyr)
library(ggplot2)

### otolith ba/ca
ggplot(data = dat[which(dat$element == "Ba_Ca"), ],
       aes(distance, value)) +
  geom_line(aes(colour = species, group = fish_id),
            show.legend = F, na.rm = T) +
  facet_grid(species ~ ., scales = "free_y") +
  xlab("Distance from core to edge of otolith (um)") +
  ylab("Ba/Ca (umol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
### otolith sr/ca
ggplot(data = dat[which(dat$element == "Sr_Ca"), ],
       aes(distance, value)) +
  geom_line(aes(colour = species, group = fish_id),
            show.legend = F, na.rm = T) +
  facet_grid(species ~ ., scales = "free_y") +
  xlab("Distance from core to edge of otolith (um)") +
  ylab("Sr/Ca (mmol/mol)") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}