Reconstructing fish histories using eye-lens isotopes

Eye lens isotopes appear to be useful for reconstructing the isotopic histories of individual fish; these histories can be compared with spatially-derived isoscapes to reconstruct individual histories for site fidelity, movement and trophic position.

bt.wallace.2014

Format

The data frame 131 × 10 contains the following columns:

fish_id	character	fish number assigned at capture

species	character	fish studied

lens_layer	integer	number assigned lamina during disection; 1 = eye lens core

d15n	numeric	bulk d15N (‰) of lamina

d13c	numeric	bulk d13C (‰) of lamina

lamina_radius	numeric	lamina radius (mm)

llamina_mid_pt	numeric	average of the two radius measurements(mm) (before and after)

lamina_weight	numeric	lamina weight (mg)

resolution	character	low or high resolution using individual lens laminae

lamina_side	character	lamina side (left or right)

Details

The dataset contains fish eye-lens stable isotopes (d13C and d15N) from the eastern Gulf of Mexico, summer 2013, including four red grouper (Epinephelus morio) and three gag (Mycteroperca microlepsis) were obtained from an FWC survey conducted 9–10 July, 2013; eight red snapper (Lutjanus campechanus) were obtained from John's Pass charter vessels on 10 September, 2013 and one white grunt (Haemulon plumierii) was obtained from a charter vessel on 11 June, 2013.

Instrument: CF-IRMS (continuous-flow isotope ratio mass spectrometry)

Reference materials: NIST857, NIST8574, NIST1577b (National Institutes of Standards and Technology)

Source

Wallace, A. A., Hollander, D. J., & Peebles, E. B. (2014). Stable isotopes in fish eye lenses as potential recorders of trophic and geographic history. PLoS One, 9(10), e108935. https://doi.org/10.1371/journal.pone.0108935

Bradley, C. J., Wallsgrove, N. J., Choy, C. A., Drazen, J. C., Hetherington, E. D., Hoen, D. K., & Popp, B. N. (2015). Trophic position estimates of marine teleosts using amino acid compound specific isotopic analysis. Limnology and oceanography: Methods, 13(9), 476-493. https://doi.org/10.1002/lom3.10041

Data availability are available at https://data.gulfresearchinitiative.org/data/R1.x135.120:0006

Traversing the paper's information via Semantic Scholar ID 39a17c40509adbe97e530196b663895a89938f6c using S2miner package

Author

Liuyong Ding, ly_ding@126.com

Concepts

eye lenses, isotopic ratios, d13C, d15N

Examples

### copy data into 'dat'
dat <- bt.wallace.2014
tibble::tibble(dat)
#> # A tibble: 131 × 10
#>    fish_id species              lens_layer  d15n  d13c lamina_radius lamina_mid_pt lamina_we…¹ resol…² lamin…³
#>      <int> <chr>                     <int> <dbl> <dbl>         <dbl>         <dbl>       <dbl> <chr>   <chr>  
#>  1      48 Lutjanus campechanus          1 10.6  -17.4            NA            NA          NA Low     NA     
#>  2      48 Lutjanus campechanus          2 11.1  -17.9            NA            NA          NA Low     NA     
#>  3      48 Lutjanus campechanus          3 10.6  -17.0            NA            NA          NA Low     NA     
#>  4      48 Lutjanus campechanus          4 10.8  -15.7            NA            NA          NA Low     NA     
#>  5      48 Lutjanus campechanus          5 11.8  -17.1            NA            NA          NA Low     NA     
#>  6      43 Lutjanus campechanus          1  9.94 -18.1            NA            NA          NA Low     NA     
#>  7      43 Lutjanus campechanus          2 10.7  -16.7            NA            NA          NA Low     NA     
#>  8      43 Lutjanus campechanus          3 11.5  -17.0            NA            NA          NA Low     NA     
#>  9      43 Lutjanus campechanus          4 12.2  -17.3            NA            NA          NA Low     NA     
#> 10      46 Lutjanus campechanus          1 10.8  -17.9            NA            NA          NA Low     NA     
#> # … with 121 more rows, and abbreviated variable names ¹​lamina_weight, ²​resolution, ³​lamina_side

if (FALSE) {
library(dplyr)
library(ggplot2)

### bulk d15N of lamina with low-resolution screening
ggplot(data = dat[which(dat$resolution == "Low"),], aes(lens_layer,d15n))+
  geom_point(aes(colour = species, group = fish_id),size = 2, show.legend = F, na.rm = T)+
  geom_line(aes(colour = species, group = fish_id),show.legend = F, na.rm = T)+
  facet_grid(species ~.,scales = "free_y")+
  xlab("Laminae")+
  ylab(expression(delta * ""^15 * "N" * " (‰ Air)"))+
  scale_x_continuous(breaks = 1:5)+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### bulk d13C of lamina with low-resolution screening
ggplot(data = dat[which(dat$resolution == "Low"),], aes(lens_layer,d13c))+
  geom_point(aes(colour = species, group = fish_id),size = 2, show.legend = F, na.rm = T)+
  geom_line(aes(colour = species, group = fish_id),show.legend = F, na.rm = T)+
  facet_grid(species ~.,scales = "free_y")+
  xlab("Laminae")+
  ylab(expression(delta * ""^13 * "C" * " (‰ PDB)"))+
  scale_x_continuous(breaks = 1:5)+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### bulk d15N of lamina with high-resolution screening
ggplot(data = dat[which(dat$resolution == "High"),], aes(lens_layer,d15n))+
  geom_point(aes(colour = species, group = fish_id),size = 2, show.legend = F, na.rm = T)+
  geom_line(aes(colour = species, group = fish_id),show.legend = F, na.rm = T)+
  facet_grid(species ~.,scales = "free_y")+
  xlab("Laminae")+
  ylab(expression(delta * ""^15 * "N" * " (‰ Air)"))+
  scale_x_continuous(breaks = 1:17)+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### bulk d13C of lamina with high-resolution screening
ggplot(data = dat[which(dat$resolution == "High"),], aes(lens_layer,d13c))+
  geom_point(aes(colour = species, group = fish_id),size = 2, show.legend = F, na.rm = T)+
  geom_line(aes(colour = species, group = fish_id),show.legend = F, na.rm = T)+
  facet_grid(species ~.,scales = "free_y")+
  xlab("Laminae")+
  ylab(expression(delta * ""^13 * "C" * " (‰ PDB)"))+
  scale_x_continuous(breaks = 1:17)+
  theme_bw() +
  theme(
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

### High-resolution comparisons of left and right eyes of white grunt specimen 2
ggplot(data = dat[-which(is.na(dat$lamina_side)),], aes(lamina_mid_pt,d13c))+
  geom_point(aes(colour = lamina_side),size = 2, na.rm = T)+
  geom_line(aes(colour = lamina_side), na.rm = T)+
  xlab("Radial midpoint of lamina (mm)")+
  ylab(expression(delta * ""^13 * "C" * " (‰ PDB)"))+
  scale_x_continuous(breaks = c(0,0.5,1,1.5,2,2.5,3))+
  theme_bw() +
  theme(
    legend.position = "top",
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )

ggplot(data = dat[-which(is.na(dat$lamina_side)),], aes(lamina_mid_pt,d15n))+
  geom_point(aes(colour = lamina_side),size = 2, na.rm = T)+
  geom_line(aes(colour = lamina_side), na.rm = T)+
  xlab("Radial midpoint of lamina (mm)")+
  ylab(expression(delta * ""^15 * "N" * " (‰ Air)"))+
  scale_x_continuous(breaks = c(0,0.5,1,1.5,2,2.5,3))+
  theme_bw() +
  theme(
    legend.position = "top",
    panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
    panel.background = element_blank(), axis.line = element_line(colour = "black"),
    text = element_text(size = 10), legend.title = element_blank(),
    plot.title = element_text(face = "bold")
  )
}